test(cli): split monolithic CLI test into per-command modules
The single run-cli.test.ts file had grown to ~5000 lines covering
every command. Split each command's tests into co-located
*.cli.test.ts files and extract shared fixtures and helpers
(writeAuthFile, runPrintedAuthLogin, readLatestLogContent, etc.)
into __tests__/helpers.ts for reuse across the new modules.
