+++ src/application/bootstrap/run-cli.test.ts
@@ -1,54 +1,17 @@
-import { Buffer } from "node:buffer";
-import { mkdir, readdir, readFile, stat } from "node:fs/promises";
+import { readdir, readFile, stat } from "node:fs/promises";
+import { join } from "node:path";
 
-import { join, relative } from "node:path";
-import { Database } from "bun:sqlite";
 import { describe, expect, test } from "bun:test";
 
-import { createCliSandbox, createTextBuffer } from "../../../__tests__/helpers.ts";
+import { createCliSandbox, createTextBuffer, readLatestLogContent } from "../../../__tests__/helpers.ts";
 import packageManifest from "../../../package.json" with { type: "json" };
-import {
-    downloadResumeSessionsTableName,
-    SqliteFileDownloadSessionStore,
-} from "../../adapters/store/sqlite-file-download-session-store.ts";
 import { resolveStorePaths } from "../../adapters/store/store-path.ts";
-import {
-    resolveBundledSkillVersionFilePath,
-    resolveCodexHomeDirectory,
-} from "../commands/skills/shared.ts";
 import { APP_NAME } from "../config/app-config.ts";
 import { CliUserError } from "../contracts/cli.ts";
 import { createTerminalColors } from "../terminal-colors.ts";
 import { executeCli } from "./run-cli.ts";
 
-const loginUrlColor = "#c09ff5";
-const searchBlockTitleColor = "#CAA8FA";
-const searchDisplayNameColor = "#59F78D";
-const defaultAuthEndpoint = "oomol.com";
-const packageName = packageManifest.name;
-const defaultSettingsFileContent = [
-    "# lang controls the CLI display language for help text, messages, and errors.",
-    "# Supported values: \"en\" (English), \"zh\" (Simplified Chinese).",
-    "# Default: auto-detect from LC_ALL, LC_MESSAGES, LANG, then system locale.",
-    "# lang = \"en\"",
-    "",
-    "# file.download.out_dir controls the default output directory used by `oo file download` when [outDir] is omitted.",
-    "# Default: ~/Downloads.",
-    "# Supported values: any non-empty path string.",
-    "# Relative values resolve from the current working directory when the command runs.",
-    "# A leading `~` expands to the current user's home directory.",
-    "# [file.download]",
-    "# out_dir = \"~/Downloads\"",
-    "",
-    "# skills.oo.allow_implicit_invocation controls whether Codex may invoke the bundled oo skill without an explicit mention.",
-    "# Supported values: true, false.",
-    "# Default: true.",
-    "# [skills.oo]",
-    "# allow_implicit_invocation = false",
-    "",
-].join("\n");
-
-describe("runCli", () => {
+describe("runCli bootstrap", () => {
     test("keeps the cli command name aligned with package metadata", () => {
         expect(APP_NAME in packageManifest.bin).toBeTrue();
     });
@@ -150,1994 +113,6 @@ describe("runCli", () => {
         }
     });
 
-    test("silently installs the managed Codex skill on the first run", async () => {
-        const sandbox = await createCliSandbox();
-        const codexHomeDirectory = resolveCodexHomeDirectory(sandbox.env);
-        const skillDirectoryPath = join(codexHomeDirectory, "skills", "oo");
-        const ownershipFilePath = join(skillDirectoryPath, "agents", "openai.yaml");
-        const versionFilePath = resolveBundledSkillVersionFilePath(skillDirectoryPath);
-
-        try {
-            await mkdir(codexHomeDirectory, { recursive: true });
-
-            const result = await sandbox.run(["--help"], {
-                version: "9.9.9",
-            });
-            const content = await readLatestLogContent(sandbox);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).not.toContain("Installed Codex skill");
-            expect(result.stderr).toBe("");
-            await expect(stat(join(skillDirectoryPath, "SKILL.md"))).resolves.toMatchObject({
-                isFile: expect.any(Function),
-            });
-            expect(await readFile(ownershipFilePath, "utf8")).toContain(
-                "allow_implicit_invocation: true",
-            );
-            expect(await readFile(versionFilePath, "utf8")).toBe("9.9.9\n");
-            expect(content).toContain(`"msg":"CLI first-run detection completed."`);
-            expect(content).toContain(`"isFirstRun":true`);
-            expect(content).toContain(`"shouldInstallMissingBundledSkills":true`);
-            expect(content).toContain(
-                `"msg":"Bundled Codex skill installed during first-run bootstrap."`,
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("does not auto-install the managed Codex skill during local development runs", async () => {
-        const sandbox = await createCliSandbox();
-        const codexHomeDirectory = resolveCodexHomeDirectory(sandbox.env);
-        const skillDirectoryPath = join(codexHomeDirectory, "skills", "oo");
-
-        try {
-            await mkdir(codexHomeDirectory, { recursive: true });
-
-            const result = await sandbox.run(["--help"]);
-            const content = await readLatestLogContent(sandbox);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).not.toContain("Installed Codex skill");
-            expect(result.stderr).toBe("");
-            await expect(stat(skillDirectoryPath)).rejects.toMatchObject({
-                code: "ENOENT",
-            });
-            expect(content).toContain(`"isFirstRun":true`);
-            expect(content).toContain(`"shouldSynchronizeBundledSkills":false`);
-            expect(content).toContain(`"shouldInstallMissingBundledSkills":false`);
-            expect(content).not.toContain(
-                `"msg":"Bundled Codex skill installed during first-run bootstrap."`,
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("does not auto-install the managed Codex skill for skills subcommands", async () => {
-        const sandbox = await createCliSandbox();
-        const codexHomeDirectory = resolveCodexHomeDirectory(sandbox.env);
-        const skillDirectoryPath = join(codexHomeDirectory, "skills", "oo");
-
-        try {
-            await mkdir(codexHomeDirectory, { recursive: true });
-
-            const result = await sandbox.run(["skills", "uninstall"]);
-
-            expect(result.exitCode).toBe(1);
-            expect(result.stdout).toBe("");
-            expect(result.stderr).toBe(
-                `Codex skill oo is not installed at ${skillDirectoryPath}.\n`,
-            );
-            await expect(stat(skillDirectoryPath)).rejects.toMatchObject({
-                code: "ENOENT",
-            });
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("synchronizes the managed Codex skill policy from persisted settings", async () => {
-        const sandbox = await createCliSandbox();
-        const codexHomeDirectory = resolveCodexHomeDirectory(sandbox.env);
-        const skillDirectoryPath = join(codexHomeDirectory, "skills", "oo");
-        const versionFilePath = resolveBundledSkillVersionFilePath(skillDirectoryPath);
-        const ownershipFilePath = join(skillDirectoryPath, "agents", "openai.yaml");
-        const settingsFilePath = join(
-            sandbox.env.XDG_CONFIG_HOME!,
-            APP_NAME,
-            "settings.toml",
-        );
-
-        try {
-            await mkdir(join(skillDirectoryPath, "agents"), { recursive: true });
-            await Bun.write(versionFilePath, "9.9.9\n");
-            await Bun.write(
-                ownershipFilePath,
-                await Bun.file(
-                    join(process.cwd(), "contrib", "skills", "oo", "agents", "openai.yaml"),
-                ).text(),
-            );
-            await Bun.write(
-                settingsFilePath,
-                [
-                    "[skills.oo]",
-                    "allow_implicit_invocation = false",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(["--help"], {
-                version: "9.9.9",
-            });
-
-            expect(result.exitCode).toBe(0);
-            expect(await readFile(ownershipFilePath, "utf8")).toContain(
-                "allow_implicit_invocation: false",
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("writes explicit skills install and uninstall logs", async () => {
-        const sandbox = await createCliSandbox();
-        const codexHomeDirectory = resolveCodexHomeDirectory(sandbox.env);
-        const skillDirectoryPath = join(codexHomeDirectory, "skills", "oo");
-
-        try {
-            await mkdir(codexHomeDirectory, { recursive: true });
-
-            const installResult = await sandbox.run(["skills", "install"], {
-                version: "9.9.9",
-            });
-            const installContent = await readLatestLogContent(sandbox);
-            const uninstallResult = await sandbox.run(["skills", "uninstall"]);
-            const uninstallContent = await readLatestLogContent(sandbox);
-
-            expect(installResult.exitCode).toBe(0);
-            expect(installContent).toContain(
-                `"msg":"Bundled Codex skill installed explicitly."`,
-            );
-            expect(installContent).toContain(`"skillName":"oo"`);
-            expect(installContent).toContain(`"path":"${skillDirectoryPath}"`);
-            expect(installContent).toContain(`"version":"9.9.9"`);
-
-            expect(uninstallResult.exitCode).toBe(0);
-            expect(uninstallContent).toContain(
-                `"msg":"Bundled Codex skill removed explicitly."`,
-            );
-            expect(uninstallContent).toContain(`"skillName":"oo"`);
-            expect(uninstallContent).toContain(`"path":"${skillDirectoryPath}"`);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("writes search request lifecycle logs", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(
-                ["search", "image processing"],
-                {
-                    fetcher: async () => new Response(JSON.stringify({
-                        packages: [],
-                    })),
-                },
-            );
-            const content = await readLatestLogContent(sandbox);
-
-            expect(result.exitCode).toBe(0);
-            expect(content).toContain(`"msg":"Search response cache miss."`);
-            expect(content).toContain(`"msg":"Search request started."`);
-            expect(content).toContain(`"msg":"Search request completed."`);
-            expect(content).toContain(`"msg":"Search response cached."`);
-            expect(content).toContain(`"path":"/v1/packages/-/intent-search"`);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("writes auth login callback logs without persisting api keys", async () => {
-        const sandbox = await createCliSandbox();
-        const encodedApiKey = Buffer.from("secret-1", "utf8").toString("base64");
-
-        try {
-            const result = await runPrintedAuthLogin(sandbox, "secret-1");
-            const content = await readLatestLogContent(sandbox);
-
-            expect(result.exitCode).toBe(0);
-            expect(content).toContain(
-                `"msg":"Auth login callback server is listening."`,
-            );
-            expect(content).toContain(`"msg":"Auth login callback received."`);
-            expect(content).toContain(
-                `"msg":"Auth login callback completed successfully."`,
-            );
-            expect(content).toContain(
-                `"msg":"Auth account persisted after browser login."`,
-            );
-            expect(content).toContain(`"hasApiKey":true`);
-            expect(content).not.toContain("secret-1");
-            expect(content).not.toContain(encodedApiKey);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("writes package info and cloud-task request logs during task creation", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(
-                ["cloud-task", "run", "foo/bar@1.2.3", "--block-id", "main"],
-                {
-                    fetcher: async (input) => {
-                        const request = toRequest(input);
-
-                        if (request.url.startsWith("https://registry.")) {
-                            return new Response(JSON.stringify({
-                                blocks: [
-                                    {
-                                        blockName: "main",
-                                        inputHandleDefs: [],
-                                        outputHandleDefs: [],
-                                        title: "Main",
-                                    },
-                                ],
-                                packageName: "foo/bar",
-                                packageVersion: "1.2.3",
-                                title: "Foo Bar",
-                            }));
-                        }
-
-                        return new Response(JSON.stringify({
-                            taskID: "task-1",
-                        }));
-                    },
-                },
-            );
-            const content = await readLatestLogContent(sandbox);
-
-            expect(result.exitCode).toBe(0);
-            expect(content).toContain(`"msg":"Package info request started."`);
-            expect(content).toContain(`"msg":"Package info request completed."`);
-            expect(content).toContain(`"msg":"Cloud task request started."`);
-            expect(content).toContain(`"msg":"Cloud task request completed."`);
-            expect(content).toContain(`"packageName":"foo/bar"`);
-            expect(content).toContain(`"packageVersion":"1.2.3"`);
-            expect(content).toContain(`"path":"/v3/users/me/tasks"`);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("writes update-check lifecycle logs when check-update finds a newer release", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            sandbox.env.npm_config_user_agent = "pnpm/10.0.0 node/v22.0.0";
-
-            const result = await sandbox.run(
-                ["check-update"],
-                {
-                    fetcher: async () => new Response(JSON.stringify({
-                        "dist-tags": {
-                            latest: "1.2.0",
-                        },
-                    })),
-                    packageName,
-                    stdout: {
-                        isTTY: true,
-                    },
-                    version: "1.0.0",
-                },
-            );
-            const content = await readLatestLogContent(sandbox);
-
-            expect(result.exitCode).toBe(0);
-            expect(content).toContain(`"msg":"CLI update check started."`);
-            expect(content).toContain(
-                `"msg":"CLI update latest-release request started."`,
-            );
-            expect(content).toContain(
-                `"msg":"CLI update latest-release request completed."`,
-            );
-            expect(content).toContain(`"msg":"CLI update notice emitted."`);
-            expect(content).toContain(`"latestVersion":"1.2.0"`);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("does not automatically check for updates after unrelated commands", async () => {
-        const sandbox = await createCliSandbox();
-        let fetchCount = 0;
-
-        try {
-            const result = await sandbox.run(
-                ["config", "path"],
-                {
-                    fetcher: async () => {
-                        fetchCount += 1;
-                        throw new Error("fetch should not be called");
-                    },
-                },
-            );
-            const content = await readLatestLogContent(sandbox);
-
-            expect(result.exitCode).toBe(0);
-            expect(fetchCount).toBe(0);
-            expect(content).not.toContain(`"msg":"CLI update check started."`);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("retries once before printing a retry-later message", async () => {
-        const sandbox = await createCliSandbox();
-        let fetchCount = 0;
-
-        try {
-            const fetcher = async () => {
-                fetchCount += 1;
-                throw new Error("temporary network failure");
-            };
-            const firstResult = await sandbox.run(
-                ["check-update"],
-                {
-                    fetcher,
-                    version: "1.0.0",
-                },
-            );
-            const secondResult = await sandbox.run(
-                ["check-update"],
-                {
-                    fetcher,
-                    version: "1.0.0",
-                },
-            );
-
-            expect(firstResult.exitCode).toBe(0);
-            expect(firstResult.stdout).toContain(
-                "Unable to check for updates right now. Please try again later.",
-            );
-            expect(firstResult.stderr).toBe("");
-            expect(secondResult.exitCode).toBe(0);
-            expect(secondResult.stdout).toContain(
-                "Unable to check for updates right now. Please try again later.",
-            );
-            expect(secondResult.stderr).toBe("");
-            expect(fetchCount).toBe(4);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("does not cache failed update checks between check-update invocations", async () => {
-        const sandbox = await createCliSandbox();
-        let fetchCount = 0;
-
-        try {
-            const firstResult = await sandbox.run(
-                ["check-update"],
-                {
-                    fetcher: async () => {
-                        fetchCount += 1;
-
-                        if (fetchCount <= 2) {
-                            throw new Error("temporary network failure");
-                        }
-
-                        return new Response(JSON.stringify({
-                            "dist-tags": {
-                                latest: "1.2.0",
-                            },
-                        }));
-                    },
-                    version: "1.0.0",
-                },
-            );
-            const secondResult = await sandbox.run(
-                ["check-update"],
-                {
-                    fetcher: async () => {
-                        fetchCount += 1;
-
-                        return new Response(JSON.stringify({
-                            "dist-tags": {
-                                latest: "1.2.0",
-                            },
-                        }));
-                    },
-                    version: "1.0.0",
-                },
-            );
-
-            expect(firstResult.exitCode).toBe(0);
-            expect(firstResult.stdout).toContain(
-                "Unable to check for updates right now. Please try again later.",
-            );
-            expect(secondResult.exitCode).toBe(0);
-            expect(secondResult.stdout).toContain("Update available 1.0.0");
-            expect(fetchCount).toBe(3);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("fetches the latest registry version on every check-update invocation", async () => {
-        const sandbox = await createCliSandbox();
-        let fetchCount = 0;
-
-        try {
-            const fetcher = async () => {
-                fetchCount += 1;
-
-                return new Response(JSON.stringify({
-                    "dist-tags": {
-                        latest: fetchCount === 1 ? "1.2.0" : "1.3.0",
-                    },
-                }));
-            };
-            const firstResult = await sandbox.run(
-                ["check-update"],
-                {
-                    fetcher,
-                    version: "1.0.0",
-                },
-            );
-            const secondResult = await sandbox.run(
-                ["check-update"],
-                {
-                    fetcher,
-                    version: "1.0.0",
-                },
-            );
-
-            expect(firstResult.exitCode).toBe(0);
-            expect(firstResult.stdout).toContain("Update available 1.0.0");
-            expect(firstResult.stdout).toContain("1.2.0");
-            expect(secondResult.exitCode).toBe(0);
-            expect(secondResult.stdout).toContain("Update available 1.0.0");
-            expect(secondResult.stdout).toContain("1.3.0");
-            expect(fetchCount).toBe(2);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("prints a friendly message when check-update receives an unsupported version", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const result = await sandbox.run(
-                ["check-update"],
-                {
-                    version: "development",
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toContain(
-                "Current version development does not support update checks.",
-            );
-            expect(result.stderr).toBe("");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("writes settings-store and config mutation logs", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const result = await sandbox.run(["config", "set", "lang", "zh"]);
-            const content = await readLatestLogContent(sandbox);
-
-            expect(result.exitCode).toBe(0);
-            expect(content).toContain(
-                `"msg":"Settings store file was missing. Initializing a default file."`,
-            );
-            expect(content).toContain(`"msg":"Settings store default file created."`);
-            expect(content).toContain(`"msg":"Settings store write completed."`);
-            expect(content).toContain(`"msg":"Config value persisted."`);
-            expect(content).toContain(`"key":"lang"`);
-            expect(content).toContain(`"value":"zh"`);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("writes auth-store and auth status logs", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(
-                ["auth", "status"],
-                {
-                    fetcher: async () => new Response(null, { status: 200 }),
-                },
-            );
-            const content = await readLatestLogContent(sandbox);
-
-            expect(result.exitCode).toBe(0);
-            expect(content).toContain(`"msg":"Auth store read completed."`);
-            expect(content).toContain(`"msg":"Current auth account resolved."`);
-            expect(content).toContain(`"msg":"Auth status request started."`);
-            expect(content).toContain(`"msg":"Auth status request completed."`);
-            expect(content).not.toContain("secret-1");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("writes sqlite cache adapter logs across repeated search invocations", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            let requestCount = 0;
-            const fetcher = async () => {
-                requestCount += 1;
-
-                return new Response(JSON.stringify({
-                    packages: [],
-                }));
-            };
-
-            const firstResult = await sandbox.run(["search", "cache me"], { fetcher });
-            const firstContent = await readLatestLogContent(sandbox);
-            const secondResult = await sandbox.run(["search", "cache me"], { fetcher });
-            const secondContent = await readLatestLogContent(sandbox);
-
-            expect(firstResult.exitCode).toBe(0);
-            expect(secondResult.exitCode).toBe(0);
-            expect(requestCount).toBe(1);
-            expect(firstContent).toContain(`"msg":"Sqlite cache namespace opened."`);
-            expect(firstContent).toContain(`"msg":"Sqlite cache lookup missed."`);
-            expect(firstContent).toContain(`"msg":"Sqlite cache value stored."`);
-            expect(secondContent).toContain(`"msg":"Sqlite cache lookup hit."`);
-            expect(secondContent).toContain(`"cacheId":"search.intent-response"`);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports file upload and persists the uploaded record", async () => {
-        const sandbox = await createCliSandbox();
-        const uploadsFilePath = resolveStorePaths({
-            appName: APP_NAME,
-            env: sandbox.env,
-            platform: process.platform,
-        }).uploadsFilePath;
-        const localFilePath = join(sandbox.env.HOME!, "sample.txt");
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-            await Bun.write(localFilePath, "hello world");
-
-            const requests: Request[] = [];
-            const result = await sandbox.run(
-                ["file", "upload", localFilePath],
-                {
-                    fetcher: async (input, init) => {
-                        const request = toRequest(input, init);
-
-                        requests.push(request);
-
-                        if (request.url.endsWith("/init")) {
-                            return new Response(JSON.stringify({
-                                data: {
-                                    part_size: 4,
-                                    presigned_urls: {
-                                        1: "https://storage.example.com/upload/1",
-                                        2: "https://storage.example.com/upload/2",
-                                        3: "https://storage.example.com/upload/3",
-                                    },
-                                    total_parts: 3,
-                                    upload_id: "upload-1",
-                                },
-                            }));
-                        }
-
-                        if (request.url.endsWith("/url")) {
-                            return new Response(JSON.stringify({
-                                data: {
-                                    expires_at: "3026-03-27T00:00:00.000Z",
-                                    file_name: "sample.txt",
-                                    file_size: 11,
-                                    mime_type: "text/plain",
-                                    url: "https://download.example.com/file-1?signature=abc",
-                                },
-                            }));
-                        }
-
-                        return new Response(null, {
-                            status: 200,
-                        });
-                    },
-                },
-            );
-            const database = new Database(uploadsFilePath, {
-                strict: true,
-            });
-
-            try {
-                expect(result.exitCode).toBe(0);
-                expect(result.stderr).toBe("");
-                expect(result.stdout).toContain("Uploaded sample.txt.");
-                expect(result.stdout).toContain(
-                    "https://download.example.com/file-1?signature=abc",
-                );
-                expect(requests.map(request => request.url)).toEqual([
-                    "https://llm.oomol.com/api/tasks/files/remote-cache/init",
-                    "https://storage.example.com/upload/1",
-                    "https://storage.example.com/upload/2",
-                    "https://storage.example.com/upload/3",
-                    "https://llm.oomol.com/api/tasks/files/remote-cache/upload-1/url",
-                ]);
-                expect(requests[0]?.headers.get("Authorization")).toBe("secret-1");
-                expect(requests[0]?.method).toBe("POST");
-                await expect(requests[0]?.json()).resolves.toEqual({
-                    file_extension: ".txt",
-                    file_name: "sample",
-                    size: 11,
-                });
-                expect(
-                    database.query(
-                        [
-                            "SELECT",
-                            "file_name AS fileName,",
-                            "file_size AS fileSize,",
-                            "download_url AS downloadUrl,",
-                            "expires_at_ms AS expiresAtMs",
-                            "FROM uploaded_files",
-                        ].join(" "),
-                    ).all(),
-                ).toEqual([
-                    {
-                        downloadUrl: "https://download.example.com/file-1?signature=abc",
-                        expiresAtMs: Date.parse("3026-03-27T00:00:00.000Z"),
-                        fileName: "sample.txt",
-                        fileSize: 11,
-                    },
-                ]);
-            }
-            finally {
-                database.close();
-            }
-        }
-        finally {
-            await Bun.file(localFilePath).delete();
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports file download, creates missing directories, and prints the labeled saved path", async () => {
-        const sandbox = await createCliSandbox();
-        const outputDirectoryPath = join(sandbox.env.HOME!, "downloads", "reports");
-        const relativeOutputDirectory = relative(sandbox.cwd, outputDirectoryPath);
-
-        try {
-            const result = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/download?id=1#fragment",
-                    relativeOutputDirectory,
-                ],
-                {
-                    fetcher: async () => {
-                        const response = new Response("hello world", {
-                            headers: {
-                                "Content-Disposition": "attachment; filename=\"report\"",
-                                "Content-Length": "11",
-                                "Content-Type": "application/pdf",
-                            },
-                            status: 200,
-                        });
-
-                        Object.defineProperty(response, "url", {
-                            value: "https://cdn.example.com/files/archive.tar.gz?signature=abc",
-                        });
-
-                        return response;
-                    },
-                    stderr: {
-                        isTTY: true,
-                    },
-                },
-            );
-            const downloadedFilePath = join(outputDirectoryPath, "report.tar.gz");
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toBe(readFileDownloadSuccessOutput(downloadedFilePath));
-            expect(result.stderr).toContain("Downloaded");
-            expect(result.stderr).toContain("100%");
-            await expect(Bun.file(downloadedFilePath).text()).resolves.toBe("hello world");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("renders file download progress with human-readable byte units", async () => {
-        const sandbox = await createCliSandbox();
-        const outputDirectoryPath = join(sandbox.env.HOME!, "downloads", "reports");
-        const relativeOutputDirectory = relative(sandbox.cwd, outputDirectoryPath);
-        const content = new Uint8Array(2 * 1024).fill(97);
-
-        try {
-            const result = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/archive.bin",
-                    relativeOutputDirectory,
-                ],
-                {
-                    fetcher: async () => new Response(content, {
-                        headers: {
-                            "Content-Disposition": "attachment; filename=\"archive.bin\"",
-                            "Content-Length": `${content.byteLength}`,
-                            "Content-Type": "application/octet-stream",
-                        },
-                        status: 200,
-                    }),
-                    stderr: {
-                        isTTY: true,
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toContain("2 KB / 2 KB (100%)");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("keeps the cursor on the line below file download progress output", async () => {
-        const sandbox = await createCliSandbox();
-        const outputDirectoryPath = join(sandbox.env.HOME!, "downloads", "reports");
-        const relativeOutputDirectory = relative(sandbox.cwd, outputDirectoryPath);
-
-        try {
-            const result = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/chunked.bin",
-                    relativeOutputDirectory,
-                ],
-                {
-                    fetcher: async () => new Response(new ReadableStream<Uint8Array>({
-                        start(controller) {
-                            controller.enqueue(new TextEncoder().encode("ab"));
-                            controller.enqueue(new TextEncoder().encode("cd"));
-                            controller.close();
-                        },
-                    }), {
-                        headers: {
-                            "Content-Disposition": "attachment; filename=\"chunked.bin\"",
-                            "Content-Length": "4",
-                            "Content-Type": "application/octet-stream",
-                        },
-                        status: 200,
-                    }),
-                    stderr: {
-                        isTTY: true,
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe(
-                "Downloading 0 B / 4 B (0%)\n"
-                + "\u001B[1A\r\u001B[2KDownloaded 4 B / 4 B (100%)\n",
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("uses the default downloads directory when file download output directory is omitted", async () => {
-        const sandbox = await createCliSandbox();
-        const outputDirectoryPath = join(sandbox.env.HOME!, "Downloads");
-
-        try {
-            const result = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/download?id=default-target",
-                ],
-                {
-                    fetcher: async () => new Response("default target", {
-                        headers: {
-                            "Content-Disposition": "attachment; filename=\"report.txt\"",
-                            "Content-Length": "14",
-                            "Content-Type": "text/plain",
-                        },
-                        status: 200,
-                    }),
-                },
-            );
-            const downloadedFilePath = join(outputDirectoryPath, "report.txt");
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toBe(readFileDownloadSuccessOutput(downloadedFilePath));
-            await expect(Bun.file(downloadedFilePath).text()).resolves.toBe(
-                "default target",
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("uses the configured file download output directory and expands a leading tilde", async () => {
-        const sandbox = await createCliSandbox();
-        const outputDirectoryPath = join(sandbox.env.HOME!, "Downloads", "reports");
-
-        try {
-            const setConfigResult = await sandbox.run([
-                "config",
-                "set",
-                "file.download.out_dir",
-                "~/Downloads/reports",
-            ]);
-            const downloadResult = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/download?id=2",
-                ],
-                {
-                    fetcher: async () => new Response("configured target", {
-                        headers: {
-                            "Content-Disposition": "attachment; filename=\"report.txt\"",
-                            "Content-Length": "17",
-                            "Content-Type": "text/plain",
-                        },
-                        status: 200,
-                    }),
-                },
-            );
-            const downloadedFilePath = join(outputDirectoryPath, "report.txt");
-
-            expect(setConfigResult.exitCode).toBe(0);
-            expect(downloadResult.exitCode).toBe(0);
-            expect(downloadResult.stdout).toBe(readFileDownloadSuccessOutput(downloadedFilePath));
-            await expect(Bun.file(downloadedFilePath).text()).resolves.toBe(
-                "configured target",
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("uses the current cli working directory when file.download.out_dir is dot", async () => {
-        const sandbox = await createCliSandbox();
-        const outputDirectoryPath = sandbox.cwd;
-
-        try {
-            const setConfigResult = await sandbox.run([
-                "config",
-                "set",
-                "file.download.out_dir",
-                ".",
-            ]);
-            const downloadResult = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/download?id=3",
-                ],
-                {
-                    fetcher: async () => new Response("cwd target", {
-                        headers: {
-                            "Content-Disposition": "attachment; filename=\"config-dot-target.txt\"",
-                            "Content-Length": "10",
-                            "Content-Type": "text/plain",
-                        },
-                        status: 200,
-                    }),
-                },
-            );
-            const downloadedFilePath = join(outputDirectoryPath, "config-dot-target.txt");
-
-            expect(setConfigResult.exitCode).toBe(0);
-            expect(downloadResult.exitCode).toBe(0);
-            expect(downloadResult.stdout).toBe(readFileDownloadSuccessOutput(downloadedFilePath));
-            await expect(Bun.file(downloadedFilePath).text()).resolves.toBe("cwd target");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("auto-renames conflicting targets and removes temporary files after file download", async () => {
-        const sandbox = await createCliSandbox();
-        const outputDirectoryPath = join(sandbox.env.HOME!, "downloads");
-
-        try {
-            await mkdir(outputDirectoryPath, { recursive: true });
-            await Bun.write(join(outputDirectoryPath, "download.pdf"), "existing");
-            await Bun.write(join(outputDirectoryPath, "download_1.oodownload"), "stale");
-
-            const result = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/download",
-                    outputDirectoryPath,
-                ],
-                {
-                    fetcher: async () => new Response("fresh pdf", {
-                        headers: {
-                            "Content-Type": "application/pdf",
-                        },
-                        status: 200,
-                    }),
-                },
-            );
-            const outputEntries = await readdir(outputDirectoryPath);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(result.stdout).toBe(
-                readFileDownloadSuccessOutput(join(outputDirectoryPath, "download_1.pdf")),
-            );
-            expect(outputEntries).toHaveLength(3);
-            expect(outputEntries).toEqual(expect.arrayContaining([
-                "download.pdf",
-                "download_1.oodownload",
-                "download_1.pdf",
-            ]));
-            await expect(Bun.file(join(outputDirectoryPath, "download_1.pdf")).text()).resolves.toBe(
-                "fresh pdf",
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("fails file download when the output path already exists as a file", async () => {
-        const sandbox = await createCliSandbox();
-        const outputPath = join(sandbox.env.HOME!, "not-a-directory");
-
-        try {
-            await Bun.write(outputPath, "occupied");
-
-            let fetchCount = 0;
-            const result = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/download",
-                    outputPath,
-                ],
-                {
-                    fetcher: async () => {
-                        fetchCount += 1;
-                        return new Response("unreachable");
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(1);
-            expect(result.stdout).toBe("");
-            expect(result.stderr).toContain("not a directory");
-            expect(fetchCount).toBe(0);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("fails file download before fetching when the url is invalid", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            let fetchCount = 0;
-            const result = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "ftp://example.com/file.txt",
-                ],
-                {
-                    fetcher: async () => {
-                        fetchCount += 1;
-                        return new Response("unreachable");
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(2);
-            expect(result.stdout).toBe("");
-            expect(result.stderr).toContain("Invalid URL");
-            expect(fetchCount).toBe(0);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("fails file download before fetching when --name or --ext is invalid", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            let fetchCount = 0;
-            const invalidNameResult = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/file.txt",
-                    "--name",
-                    "..",
-                ],
-                {
-                    fetcher: async () => {
-                        fetchCount += 1;
-                        return new Response("unreachable");
-                    },
-                },
-            );
-            const invalidExtResult = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/file.txt",
-                    "--ext",
-                    "..txt",
-                ],
-                {
-                    fetcher: async () => {
-                        fetchCount += 1;
-                        return new Response("unreachable");
-                    },
-                },
-            );
-
-            expect(invalidNameResult.exitCode).toBe(2);
-            expect(invalidNameResult.stdout).toBe("");
-            expect(invalidNameResult.stderr).toContain("Invalid value for --name");
-
-            expect(invalidExtResult.exitCode).toBe(2);
-            expect(invalidExtResult.stdout).toBe("");
-            expect(invalidExtResult.stderr).toContain("Invalid value for --ext");
-            expect(fetchCount).toBe(0);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("fails file download when the download request returns a non-success status", async () => {
-        const sandbox = await createCliSandbox();
-        const outputDirectoryPath = join(sandbox.env.HOME!, "downloads");
-
-        try {
-            await mkdir(outputDirectoryPath, { recursive: true });
-
-            const result = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/missing.txt",
-                    outputDirectoryPath,
-                ],
-                {
-                    fetcher: async () => new Response("missing", {
-                        status: 404,
-                    }),
-                },
-            );
-
-            expect(result.exitCode).toBe(1);
-            expect(result.stdout).toBe("");
-            expect(result.stderr).toContain("HTTP 404");
-            expect(await readdir(outputDirectoryPath)).toEqual([]);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("fails file download when the output directory cannot be created", async () => {
-        const sandbox = await createCliSandbox();
-        const occupiedPath = join(sandbox.env.HOME!, "occupied");
-        const nestedOutputDirectory = join(occupiedPath, "nested");
-
-        try {
-            await Bun.write(occupiedPath, "occupied");
-
-            let fetchCount = 0;
-            const result = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/file.txt",
-                    nestedOutputDirectory,
-                ],
-                {
-                    fetcher: async () => {
-                        fetchCount += 1;
-                        return new Response("unreachable");
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(1);
-            expect(result.stdout).toBe("");
-            expect(result.stderr).toContain("Failed to prepare the output directory");
-            expect(fetchCount).toBe(0);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("uses the default download file name when the final response url has no file segment", async () => {
-        const sandbox = await createCliSandbox();
-        const outputDirectoryPath = join(sandbox.env.HOME!, "downloads");
-
-        try {
-            const result = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/files/",
-                    outputDirectoryPath,
-                ],
-                {
-                    fetcher: async () => {
-                        const response = new Response("payload", {
-                            status: 200,
-                        });
-
-                        Object.defineProperty(response, "url", {
-                            value: "https://example.com/files/",
-                        });
-
-                        return response;
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(result.stdout).toBe(
-                readFileDownloadSuccessOutput(join(outputDirectoryPath, "download")),
-            );
-            await expect(Bun.file(join(outputDirectoryPath, "download")).text()).resolves.toBe(
-                "payload",
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("preserves the temporary file and resume session when file download fails mid-download", async () => {
-        const sandbox = await createCliSandbox();
-        const outputDirectoryPath = join(sandbox.env.HOME!, "downloads");
-        const downloadSessionsFilePath = resolveStorePaths({
-            appName: APP_NAME,
-            env: sandbox.env,
-            platform: process.platform,
-        }).downloadSessionsFilePath;
-
-        try {
-            await mkdir(outputDirectoryPath, { recursive: true });
-
-            const result = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/broken.txt",
-                    outputDirectoryPath,
-                ],
-                {
-                    fetcher: async () => {
-                        let emittedChunk = false;
-                        const body = new ReadableStream<Uint8Array>({
-                            pull(controller) {
-                                if (!emittedChunk) {
-                                    emittedChunk = true;
-                                    controller.enqueue(new TextEncoder().encode("partial"));
-                                    return;
-                                }
-
-                                controller.error(new Error("Connection dropped."));
-                            },
-                        });
-                        const response = new Response(body, {
-                            headers: {
-                                "Content-Disposition": "attachment; filename=\"broken.txt\"",
-                            },
-                            status: 200,
-                        });
-
-                        Object.defineProperty(response, "url", {
-                            value: "https://example.com/broken.txt",
-                        });
-
-                        return response;
-                    },
-                    stderr: {
-                        isTTY: true,
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(1);
-            expect(result.stdout).toBe("");
-            expect(result.stderr).toContain("Downloading");
-            expect(result.stderr).not.toContain("Downloaded");
-            expect(result.stderr).toContain("Connection dropped.");
-            const outputEntries = await readdir(outputDirectoryPath);
-
-            expect(outputEntries).toHaveLength(1);
-            expect(outputEntries).toEqual([
-                "broken.oodownload",
-            ]);
-            expect(countDownloadResumeSessions(downloadSessionsFilePath)).toBe(1);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("cleans up download resume sessions older than 14 days when file download starts", async () => {
-        const sandbox = await createCliSandbox();
-        const outputDirectoryPath = join(sandbox.env.HOME!, "downloads");
-        const downloadSessionsFilePath = resolveStorePaths({
-            appName: APP_NAME,
-            env: sandbox.env,
-            platform: process.platform,
-        }).downloadSessionsFilePath;
-        const sessionStore = new SqliteFileDownloadSessionStore(
-            downloadSessionsFilePath,
-        );
-
-        try {
-            await mkdir(outputDirectoryPath, { recursive: true });
-            sessionStore.saveDownloadSession({
-                entityTag: "\"stale-1\"",
-                finalUrl: "https://cdn.example.com/stale.txt",
-                id: "0195f5fe-ec30-7000-8000-000000000011",
-                lastModified: "Wed, 01 Jan 2025 00:00:00 GMT",
-                outDirPath: outputDirectoryPath,
-                requestUrl: "https://example.com/stale.txt",
-                requestedExtension: "",
-                requestedName: "",
-                resolvedBaseName: "stale",
-                resolvedExtension: "txt",
-                tempFileName: "stale.oodownload",
-                totalBytes: 64,
-                updatedAtMs: Date.now() - (15 * 24 * 60 * 60 * 1000),
-            });
-
-            const result = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/fresh.txt",
-                    outputDirectoryPath,
-                ],
-                {
-                    fetcher: async () => new Response("fresh", {
-                        headers: {
-                            "Content-Disposition": "attachment; filename=\"fresh.txt\"",
-                            "Content-Length": "5",
-                        },
-                        status: 200,
-                    }),
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(countDownloadResumeSessions(downloadSessionsFilePath)).toBe(0);
-        }
-        finally {
-            sessionStore.close();
-            await sandbox.cleanup();
-        }
-    });
-
-    test("resumes file download with HTTP Range after a mid-download failure", async () => {
-        const sandbox = await createCliSandbox();
-        const outputDirectoryPath = join(sandbox.env.HOME!, "downloads");
-        const downloadSessionsFilePath = resolveStorePaths({
-            appName: APP_NAME,
-            env: sandbox.env,
-            platform: process.platform,
-        }).downloadSessionsFilePath;
-        let requestCount = 0;
-
-        try {
-            await mkdir(outputDirectoryPath, { recursive: true });
-
-            const fetcher = async (_input: string | URL | Request, init?: RequestInit) => {
-                requestCount += 1;
-                const headers = new Headers(init?.headers);
-
-                if (requestCount === 1) {
-                    expect(headers.get("Range")).toBeNull();
-
-                    let emittedChunk = false;
-                    const body = new ReadableStream<Uint8Array>({
-                        pull(controller) {
-                            if (!emittedChunk) {
-                                emittedChunk = true;
-                                controller.enqueue(new TextEncoder().encode("partial"));
-                                return;
-                            }
-
-                            controller.error(new Error("Connection dropped."));
-                        },
-                    });
-                    const response = new Response(body, {
-                        headers: {
-                            "Content-Disposition": "attachment; filename=\"broken.txt\"",
-                            "Content-Length": "15",
-                            "ETag": "\"resume-1\"",
-                            "Last-Modified": "Wed, 01 Jan 2025 00:00:00 GMT",
-                        },
-                        status: 200,
-                    });
-
-                    Object.defineProperty(response, "url", {
-                        value: "https://example.com/broken.txt",
-                    });
-
-                    return response;
-                }
-
-                expect(headers.get("Range")).toBe("bytes=7-");
-                expect(headers.get("If-Range")).toBe("\"resume-1\"");
-
-                const response = new Response("-payload", {
-                    headers: {
-                        "Content-Length": "8",
-                        "Content-Range": "bytes 7-14/15",
-                        "ETag": "\"resume-1\"",
-                        "Last-Modified": "Wed, 01 Jan 2025 00:00:00 GMT",
-                    },
-                    status: 206,
-                });
-
-                Object.defineProperty(response, "url", {
-                    value: "https://example.com/broken.txt",
-                });
-
-                return response;
-            };
-
-            const firstResult = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/broken.txt",
-                    outputDirectoryPath,
-                ],
-                {
-                    fetcher,
-                },
-            );
-            const secondResult = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/broken.txt",
-                    outputDirectoryPath,
-                ],
-                {
-                    fetcher,
-                },
-            );
-            const downloadedFilePath = join(outputDirectoryPath, "broken.txt");
-
-            expect(firstResult.exitCode).toBe(1);
-            expect(firstResult.stderr).toContain("Connection dropped.");
-            expect(secondResult.exitCode).toBe(0);
-            expect(secondResult.stderr).toBe("");
-            expect(secondResult.stdout).toBe(readFileDownloadSuccessOutput(downloadedFilePath));
-            expect(requestCount).toBe(2);
-            await expect(Bun.file(downloadedFilePath).text()).resolves.toBe("partial-payload");
-            expect(await readdir(outputDirectoryPath)).toEqual([
-                "broken.txt",
-            ]);
-            expect(countDownloadResumeSessions(downloadSessionsFilePath)).toBe(0);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("restarts file download from the beginning when the server ignores Range", async () => {
-        const sandbox = await createCliSandbox();
-        const outputDirectoryPath = join(sandbox.env.HOME!, "downloads");
-        const downloadSessionsFilePath = resolveStorePaths({
-            appName: APP_NAME,
-            env: sandbox.env,
-            platform: process.platform,
-        }).downloadSessionsFilePath;
-        let requestCount = 0;
-
-        try {
-            await mkdir(outputDirectoryPath, { recursive: true });
-
-            const fetcher = async (_input: string | URL | Request, init?: RequestInit) => {
-                requestCount += 1;
-                const headers = new Headers(init?.headers);
-
-                if (requestCount === 1) {
-                    expect(headers.get("Range")).toBeNull();
-
-                    let emittedChunk = false;
-                    const body = new ReadableStream<Uint8Array>({
-                        pull(controller) {
-                            if (!emittedChunk) {
-                                emittedChunk = true;
-                                controller.enqueue(new TextEncoder().encode("stale"));
-                                return;
-                            }
-
-                            controller.error(new Error("Connection dropped."));
-                        },
-                    });
-                    const response = new Response(body, {
-                        headers: {
-                            "Content-Disposition": "attachment; filename=\"restart.txt\"",
-                            "Content-Length": "13",
-                            "ETag": "\"restart-1\"",
-                        },
-                        status: 200,
-                    });
-
-                    Object.defineProperty(response, "url", {
-                        value: "https://example.com/restart.txt",
-                    });
-
-                    return response;
-                }
-
-                expect(headers.get("Range")).toBe("bytes=5-");
-                expect(headers.get("If-Range")).toBe("\"restart-1\"");
-
-                const response = new Response("stale-payload", {
-                    headers: {
-                        "Content-Disposition": "attachment; filename=\"restart.txt\"",
-                        "Content-Length": "13",
-                        "ETag": "\"restart-2\"",
-                    },
-                    status: 200,
-                });
-
-                Object.defineProperty(response, "url", {
-                    value: "https://example.com/restart.txt",
-                });
-
-                return response;
-            };
-
-            const firstResult = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/restart.txt",
-                    outputDirectoryPath,
-                ],
-                {
-                    fetcher,
-                },
-            );
-            const secondResult = await sandbox.run(
-                [
-                    "file",
-                    "download",
-                    "https://example.com/restart.txt",
-                    outputDirectoryPath,
-                ],
-                {
-                    fetcher,
-                },
-            );
-            const downloadedFilePath = join(outputDirectoryPath, "restart.txt");
-
-            expect(firstResult.exitCode).toBe(1);
-            expect(secondResult.exitCode).toBe(0);
-            expect(secondResult.stderr).toBe("");
-            expect(secondResult.stdout).toBe(readFileDownloadSuccessOutput(downloadedFilePath));
-            expect(requestCount).toBe(2);
-            await expect(Bun.file(downloadedFilePath).text()).resolves.toBe("stale-payload");
-            expect(await readdir(outputDirectoryPath)).toEqual([
-                "restart.txt",
-            ]);
-            expect(countDownloadResumeSessions(downloadSessionsFilePath)).toBe(0);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports file upload json output for both --json and --format=json", async () => {
-        const sandbox = await createCliSandbox();
-        const firstFilePath = join(sandbox.env.HOME!, "sample-json.txt");
-        const secondFilePath = join(sandbox.env.HOME!, "sample-format.txt");
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-            await Bun.write(firstFilePath, "json upload");
-            await Bun.write(secondFilePath, "format upload");
-
-            let uploadIndex = 0;
-            const fetcher = async (input: string | URL | Request, init?: RequestInit) => {
-                const request = toRequest(input, init);
-
-                if (request.url.endsWith("/init")) {
-                    uploadIndex += 1;
-
-                    return new Response(JSON.stringify({
-                        data: {
-                            part_size: 32,
-                            presigned_urls: {
-                                1: `https://storage.example.com/upload/${uploadIndex}`,
-                            },
-                            total_parts: 1,
-                            upload_id: `upload-${uploadIndex}`,
-                        },
-                    }));
-                }
-
-                if (request.url.endsWith("/url")) {
-                    return new Response(JSON.stringify({
-                        data: {
-                            expires_at: "3026-03-27T00:00:00.000Z",
-                            file_name: "sample.txt",
-                            file_size: 11,
-                            mime_type: "text/plain",
-                            url: `https://download.example.com/file-${uploadIndex}?signature=abc`,
-                        },
-                    }));
-                }
-
-                return new Response(null, {
-                    status: 200,
-                });
-            };
-
-            const jsonAliasResult = await sandbox.run(
-                ["file", "upload", firstFilePath, "--json"],
-                {
-                    fetcher,
-                },
-            );
-            const jsonFormatResult = await sandbox.run(
-                ["file", "upload", secondFilePath, "--format=json"],
-                {
-                    fetcher,
-                },
-            );
-
-            expect(jsonAliasResult.exitCode).toBe(0);
-            expect(jsonAliasResult.stderr).toBe("");
-            expect(JSON.parse(jsonAliasResult.stdout)).toMatchObject({
-                downloadUrl: "https://download.example.com/file-1?signature=abc",
-                expiresAt: "3026-03-27T00:00:00.000Z",
-                fileName: "sample-json.txt",
-                fileSize: 11,
-                status: "active",
-            });
-            expect(jsonFormatResult.exitCode).toBe(0);
-            expect(jsonFormatResult.stderr).toBe("");
-            expect(JSON.parse(jsonFormatResult.stdout)).toMatchObject({
-                downloadUrl: "https://download.example.com/file-2?signature=abc",
-                expiresAt: "3026-03-27T00:00:00.000Z",
-                fileName: "sample-format.txt",
-                fileSize: 13,
-                status: "active",
-            });
-        }
-        finally {
-            await Bun.file(firstFilePath).delete();
-            await Bun.file(secondFilePath).delete();
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports file list filters and cleanup json output", async () => {
-        const sandbox = await createCliSandbox();
-        const uploadsFilePath = resolveStorePaths({
-            appName: APP_NAME,
-            env: sandbox.env,
-            platform: process.platform,
-        }).uploadsFilePath;
-
-        try {
-            await mkdir(
-                join(
-                    sandbox.env.XDG_CONFIG_HOME!,
-                    APP_NAME,
-                    "data",
-                ),
-                {
-                    recursive: true,
-                },
-            );
-
-            const database = new Database(uploadsFilePath, {
-                create: true,
-                strict: true,
-            });
-
-            try {
-                database.run(
-                    [
-                        "CREATE TABLE uploaded_files (",
-                        "id TEXT PRIMARY KEY NOT NULL,",
-                        "file_name TEXT NOT NULL,",
-                        "file_size INTEGER NOT NULL,",
-                        "download_url TEXT NOT NULL,",
-                        "uploaded_at_ms INTEGER NOT NULL,",
-                        "expires_at_ms INTEGER NOT NULL",
-                        ") STRICT",
-                    ].join(" "),
-                );
-                database.run(
-                    [
-                        "INSERT INTO uploaded_files (",
-                        "id, file_name, file_size, download_url, uploaded_at_ms, expires_at_ms",
-                        ") VALUES",
-                        "('0195f5fe-ec27-7000-8000-000000000008', 'expired.txt', 10, 'https://download.example.com/expired', 1000, 2000),",
-                        "('0195f5fe-ec28-7000-8000-000000000009', 'active.txt', 11, 'https://download.example.com/active', 3000, 32503680000000)",
-                    ].join(" "),
-                );
-            }
-            finally {
-                database.close();
-            }
-
-            const listTextResult = await sandbox.run(
-                ["file", "list", "--status=active", "--limit=1"],
-            );
-            const listJsonResult = await sandbox.run(
-                ["file", "list", "--status=expired", "--json"],
-            );
-            const cleanupResult = await sandbox.run(
-                ["file", "cleanup", "--json"],
-            );
-            const listAfterCleanup = await sandbox.run(
-                ["file", "list", "--json"],
-            );
-
-            expect(listTextResult.exitCode).toBe(0);
-            expect(listTextResult.stderr).toBe("");
-            expect(listTextResult.stdout).toContain("active.txt");
-            expect(listTextResult.stdout).toContain("https://download.example.com/active");
-
-            expect(listJsonResult.exitCode).toBe(0);
-            expect(JSON.parse(listJsonResult.stdout)).toEqual([
-                {
-                    downloadUrl: "https://download.example.com/expired",
-                    expiresAt: new Date(2000).toISOString(),
-                    fileName: "expired.txt",
-                    fileSize: 10,
-                    id: "0195f5fe-ec27-7000-8000-000000000008",
-                    status: "expired",
-                    uploadedAt: new Date(1000).toISOString(),
-                },
-            ]);
-
-            expect(cleanupResult.exitCode).toBe(0);
-            expect(JSON.parse(cleanupResult.stdout)).toEqual({
-                deletedCount: 1,
-            });
-            expect(listAfterCleanup.exitCode).toBe(0);
-            expect(JSON.parse(listAfterCleanup.stdout)).toEqual([
-                {
-                    downloadUrl: "https://download.example.com/active",
-                    expiresAt: new Date(32503680000000).toISOString(),
-                    fileName: "active.txt",
-                    fileSize: 11,
-                    id: "0195f5fe-ec28-7000-8000-000000000009",
-                    status: "active",
-                    uploadedAt: new Date(3000).toISOString(),
-                },
-            ]);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("continues search when the sqlite cache database is locked", async () => {
-        const sandbox = await createCliSandbox();
-        const cacheFilePath = resolveStorePaths({
-            appName: APP_NAME,
-            env: sandbox.env,
-            platform: process.platform,
-        }).cacheFilePath;
-        let lockDatabase: Database | undefined;
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-            const dataDirectoryPath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "data",
-            );
-
-            await mkdir(dataDirectoryPath, { recursive: true });
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            lockDatabase = new Database(cacheFilePath, {
-                create: true,
-                strict: true,
-            });
-            lockDatabase.run("PRAGMA journal_mode = WAL;");
-            lockDatabase.run("BEGIN IMMEDIATE;");
-
-            let requestCount = 0;
-            const result = await sandbox.run(
-                ["search", "cache lock"],
-                {
-                    fetcher: async () => {
-                        requestCount += 1;
-
-                        return new Response(JSON.stringify({
-                            packages: [],
-                        }));
-                    },
-                },
-            );
-            const content = await readLatestLogContent(sandbox);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(requestCount).toBe(1);
-            expect(content).toContain(`"category":"recoverable_cache"`);
-            expect(content).toContain(
-                `"msg":"Sqlite cache namespace is temporarily unavailable because the database is locked."`,
-            );
-        }
-        finally {
-            if (lockDatabase) {
-                lockDatabase.run("ROLLBACK;");
-                lockDatabase.close();
-            }
-            await sandbox.cleanup();
-        }
-    });
-
-    test("continues search when the sqlite cache database file cannot be opened", async () => {
-        const sandbox = await createCliSandbox();
-        const cacheFilePath = resolveStorePaths({
-            appName: APP_NAME,
-            env: sandbox.env,
-            platform: process.platform,
-        }).cacheFilePath;
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-            const dataDirectoryPath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "data",
-            );
-
-            await mkdir(dataDirectoryPath, { recursive: true });
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-            await mkdir(cacheFilePath, { recursive: true });
-
-            let requestCount = 0;
-            const result = await sandbox.run(
-                ["search", "cache unavailable", "--json"],
-                {
-                    fetcher: async () => {
-                        requestCount += 1;
-
-                        return new Response(JSON.stringify({
-                            packages: [],
-                        }));
-                    },
-                },
-            );
-            const content = await readLatestLogContent(sandbox);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toBe("[]\n");
-            expect(result.stderr).toBe("");
-            expect(requestCount).toBe(1);
-            expect(content).toContain(`"category":"recoverable_cache"`);
-            expect(content).toContain(`"sqliteErrorCode":"SQLITE_CANTOPEN"`);
-            expect(content).toContain(`"storePathKind":"directory"`);
-            expect(content).toContain(`"storePathExists":true`);
-            expect(content).toContain(`"parentDirectoryExists":true`);
-            expect(content).toContain(`"parentDirectoryWritable":true`);
-            expect(content).toContain(
-                `"msg":"Sqlite cache namespace is temporarily unavailable because the database file cannot be opened."`,
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
     test("prints the current log file path to stderr when --debug is set", async () => {
         const sandbox = await createCliSandbox();
 
@@ -2153,77 +128,6 @@ describe("runCli", () => {
 
             expect(result.exitCode).toBe(0);
             expect(result.stderr).toBe(`${logFilePath}\n`);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("prints the resolved log directory path", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const logDirectoryPath = resolveStorePaths({
-                appName: APP_NAME,
-                env: sandbox.env,
-                platform: process.platform,
-            }).logDirectoryPath;
-            const result = await sandbox.run(["log", "path"]);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toBe(`${logDirectoryPath}\n`);
-            expect(result.stderr).toBe("");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("prints the previous log file by default", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const logDirectoryPath = resolveStorePaths({
-                appName: APP_NAME,
-                env: sandbox.env,
-                platform: process.platform,
-            }).logDirectoryPath;
-
-            await mkdir(logDirectoryPath, { recursive: true });
-            await Bun.write(join(logDirectoryPath, "debug-0001.log"), "first-log");
-            await Bun.write(join(logDirectoryPath, "debug-0002.log"), "second-log");
-
-            const result = await sandbox.run(["log", "print"]);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toBe("second-log\n");
-            expect(result.stderr).toBe("");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("prints a selected previous log file by index", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const logDirectoryPath = resolveStorePaths({
-                appName: APP_NAME,
-                env: sandbox.env,
-                platform: process.platform,
-            }).logDirectoryPath;
-
-            await mkdir(logDirectoryPath, { recursive: true });
-            await Bun.write(join(logDirectoryPath, "debug-0001.log"), "first-log");
-            await Bun.write(join(logDirectoryPath, "debug-0002.log"), "second-log");
-            await Bun.write(join(logDirectoryPath, "debug-0003.log"), "third-log");
-
-            const result = await sandbox.run(["log", "print", "2"]);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toBe("second-log\n");
-            expect(result.stderr).toBe("");
         }
         finally {
             await sandbox.cleanup();
@@ -2270,38 +174,6 @@ describe("runCli", () => {
         }
     });
 
-    test("renders alias descriptions in login and logout help", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const loginHelp = await sandbox.run(["login", "--help"]);
-            const logoutHelp = await sandbox.run(["logout", "--help"]);
-            const loginDescriptionIndex = loginHelp.stdout.indexOf(
-                "Log in with an OOMOL account in the browser.",
-            );
-            const loginAliasIndex = loginHelp.stdout.indexOf(
-                "Alias for auth login.",
-            );
-            const logoutDescriptionIndex = logoutHelp.stdout.indexOf(
-                "Remove the current account from persisted auth data.",
-            );
-            const logoutAliasIndex = logoutHelp.stdout.indexOf(
-                "Alias for auth logout.",
-            );
-
-            expect(loginHelp.exitCode).toBe(0);
-            expect(loginDescriptionIndex).toBeGreaterThanOrEqual(0);
-            expect(loginAliasIndex).toBeGreaterThan(loginDescriptionIndex);
-
-            expect(logoutHelp.exitCode).toBe(0);
-            expect(logoutDescriptionIndex).toBeGreaterThanOrEqual(0);
-            expect(logoutAliasIndex).toBeGreaterThan(logoutDescriptionIndex);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
     test("renders branded colors in help when stdout supports colors", async () => {
         const sandbox = await createCliSandbox();
         const colors = createTerminalColors(true);
@@ -2326,2748 +198,18 @@ describe("runCli", () => {
         }
     });
 
-    test("renders supported shells in completion help", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const result = await sandbox.run(["completion", "--help"]);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toContain("Target shell");
-            expect(result.stdout).toContain("\"bash\"");
-            expect(result.stdout).toContain("\"zsh\"");
-            expect(result.stdout).toContain("\"fish\"");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("renders localized choices metadata in Chinese completion help", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const result = await sandbox.run(["--lang", "zh", "completion", "--help"]);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toContain("目标 shell");
-            expect(result.stdout).toContain("(可选值: \"bash\", \"zsh\", \"fish\")");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("persists the configured locale and allows explicit override", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const setResult = await sandbox.run(["config", "set", "lang", "zh"]);
-            const persistedHelp = await sandbox.run(["--help"]);
-            const overriddenHelp = await sandbox.run(["--lang", "en", "--help"]);
-
-            expect(setResult.exitCode).toBe(0);
-            expect(setResult.stdout).toContain("Set lang to zh.");
-            expect(persistedHelp.stdout).not.toContain("用法：");
-            expect(overriddenHelp.stdout).not.toContain("Usage:");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports config path list get set and unset", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const configPathResult = await sandbox.run(["config", "path"]);
-            const listBeforeSetResult = await sandbox.run(["config", "list"]);
-            const setResult = await sandbox.run(["config", "set", "lang", "zh"]);
-            const listAfterSetResult = await sandbox.run(["config", "list"]);
-            const getResult = await sandbox.run(["config", "get", "lang"]);
-            const unsetResult = await sandbox.run(["config", "unset", "lang"]);
-            const listAfterUnsetResult = await sandbox.run(["config", "list"]);
-            const getAfterUnsetResult = await sandbox.run(["config", "get", "lang"]);
-
-            expect(configPathResult.exitCode).toBe(0);
-            expect(configPathResult.stdout).toBe(
-                `${join(sandbox.env.XDG_CONFIG_HOME!, APP_NAME, "settings.toml")}\n`,
-            );
-            expect(listBeforeSetResult.exitCode).toBe(0);
-            expect(listBeforeSetResult.stdout).toBe("");
-            expect(setResult.exitCode).toBe(0);
-            expect(listAfterSetResult.exitCode).toBe(0);
-            expect(listAfterSetResult.stdout).toBe("lang=zh\n");
-            expect(getResult.stdout).toBe("zh\n");
-            expect(unsetResult.exitCode).toBe(0);
-            expect(listAfterUnsetResult.exitCode).toBe(0);
-            expect(listAfterUnsetResult.stdout).toBe("");
-            expect(getAfterUnsetResult.stdout).toBe("");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports the oo skill implicit invocation config key", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const setResult = await sandbox.run([
-                "config",
-                "set",
-                "skills.oo.allow_implicit_invocation",
-                "false",
-            ]);
-            const listResult = await sandbox.run(["config", "list"]);
-            const getResult = await sandbox.run([
-                "config",
-                "get",
-                "skills.oo.allow_implicit_invocation",
-            ]);
-            const unsetResult = await sandbox.run([
-                "config",
-                "unset",
-                "skills.oo.allow_implicit_invocation",
-            ]);
-            const getAfterUnsetResult = await sandbox.run([
-                "config",
-                "get",
-                "skills.oo.allow_implicit_invocation",
-            ]);
-
-            expect(setResult.exitCode).toBe(0);
-            expect(setResult.stdout).toBe(
-                "Set skills.oo.allow_implicit_invocation to false.\n",
-            );
-            expect(listResult.stdout).toBe(
-                "skills.oo.allow_implicit_invocation=false\n",
-            );
-            expect(getResult.stdout).toBe("false\n");
-            expect(unsetResult.exitCode).toBe(0);
-            expect(getAfterUnsetResult.stdout).toBe("");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports the file download output directory config key", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const setResult = await sandbox.run([
-                "config",
-                "set",
-                "file.download.out_dir",
-                "~/Downloads/reports",
-            ]);
-            const listResult = await sandbox.run(["config", "list"]);
-            const getResult = await sandbox.run([
-                "config",
-                "get",
-                "file.download.out_dir",
-            ]);
-            const unsetResult = await sandbox.run([
-                "config",
-                "unset",
-                "file.download.out_dir",
-            ]);
-            const getAfterUnsetResult = await sandbox.run([
-                "config",
-                "get",
-                "file.download.out_dir",
-            ]);
-
-            expect(setResult.exitCode).toBe(0);
-            expect(setResult.stdout).toBe(
-                "Set file.download.out_dir to ~/Downloads/reports.\n",
-            );
-            expect(listResult.stdout).toBe(
-                "file.download.out_dir=~/Downloads/reports\n",
-            );
-            expect(getResult.stdout).toBe("~/Downloads/reports\n");
-            expect(unsetResult.exitCode).toBe(0);
-            expect(getAfterUnsetResult.stdout).toBe("");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("config set immediately synchronizes the installed managed skill policy", async () => {
-        const sandbox = await createCliSandbox();
-        const codexHomeDirectory = resolveCodexHomeDirectory(sandbox.env);
-        const skillDirectoryPath = join(codexHomeDirectory, "skills", "oo");
-        const ownershipFilePath = join(skillDirectoryPath, "agents", "openai.yaml");
-        const versionFilePath = resolveBundledSkillVersionFilePath(skillDirectoryPath);
-
-        try {
-            await mkdir(join(skillDirectoryPath, "agents"), { recursive: true });
-            await Bun.write(versionFilePath, "9.9.9\n");
-            await Bun.write(
-                ownershipFilePath,
-                await Bun.file(
-                    join(process.cwd(), "contrib", "skills", "oo", "agents", "openai.yaml"),
-                ).text(),
-            );
-
-            const result = await sandbox.run([
-                "config",
-                "set",
-                "skills.oo.allow_implicit_invocation",
-                "false",
-            ], {
-                version: "9.9.9",
-            });
-
-            expect(result.exitCode).toBe(0);
-            expect(await readFile(ownershipFilePath, "utf8")).toContain(
-                "allow_implicit_invocation: false",
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("config unset immediately restores the installed managed skill policy default", async () => {
-        const sandbox = await createCliSandbox();
-        const codexHomeDirectory = resolveCodexHomeDirectory(sandbox.env);
-        const skillDirectoryPath = join(codexHomeDirectory, "skills", "oo");
-        const ownershipFilePath = join(skillDirectoryPath, "agents", "openai.yaml");
-        const versionFilePath = resolveBundledSkillVersionFilePath(skillDirectoryPath);
-        const settingsFilePath = join(
-            sandbox.env.XDG_CONFIG_HOME!,
-            APP_NAME,
-            "settings.toml",
-        );
-
-        try {
-            await mkdir(join(skillDirectoryPath, "agents"), { recursive: true });
-            await Bun.write(versionFilePath, "9.9.9\n");
-            await Bun.write(
-                ownershipFilePath,
-                [
-                    "interface:",
-                    "  display_name: oo",
-                    "  short_description: Check OOMOL first for ready-made capabilities",
-                    "  default_prompt: \"Use $oo when I ask for a ready-made capability.\"",
-                    "",
-                    "policy:",
-                    "  allow_implicit_invocation: false",
-                    "",
-                    "# OOMOL",
-                    "",
-                ].join("\n"),
-            );
-            await Bun.write(
-                settingsFilePath,
-                [
-                    "[skills.oo]",
-                    "allow_implicit_invocation = false",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run([
-                "config",
-                "unset",
-                "skills.oo.allow_implicit_invocation",
-            ], {
-                version: "9.9.9",
-            });
-
-            expect(result.exitCode).toBe(0);
-            expect(await readFile(ownershipFilePath, "utf8")).toContain(
-                "allow_implicit_invocation: true",
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("renders config list help with configured wording", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const englishConfigHelp = await sandbox.run(["config", "--help"]);
-            const englishListHelp = await sandbox.run(["config", "list", "--help"]);
-            const chineseConfigHelp = await sandbox.run(["--lang", "zh", "config", "--help"]);
-            const chineseListHelp = await sandbox.run(["--lang", "zh", "config", "list", "--help"]);
-
-            expect(englishConfigHelp.exitCode).toBe(0);
-            expect(englishConfigHelp.stdout).toContain("List configured values");
-            expect(englishConfigHelp.stdout).toContain("Show config file path");
-
-            expect(englishListHelp.exitCode).toBe(0);
-            expect(englishListHelp.stdout).toContain(
-                "Print all persisted configuration values that are currently configured.",
-            );
-
-            expect(chineseConfigHelp.exitCode).toBe(0);
-            expect(chineseConfigHelp.stdout).toContain("查看已配置的配置值");
-            expect(chineseConfigHelp.stdout).toContain("显示配置文件路径");
-
-            expect(chineseListHelp.exitCode).toBe(0);
-            expect(chineseListHelp.stdout).toContain("查看当前已配置的全部持久化配置值。");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports auth login and updates the existing account without duplication", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-            const firstLogin = await runPrintedAuthLogin(sandbox, "secret-1");
-            const secondLogin = await runPrintedAuthLogin(sandbox, "secret-2");
-            const authFileContent = await readFile(authFilePath, "utf8");
-
-            expect(firstLogin.exitCode).toBe(0);
-            expect(firstLogin.stdout).toContain(
-                "Open this URL in your browser to continue:",
-            );
-            expect(firstLogin.stdout).toContain(
-                readAuthLoginUrlPrefix(defaultAuthEndpoint),
-            );
-            expect(firstLogin.stdout).toContain("✓ Logged in to oomol.com account Alice");
-            expect(firstLogin.stdout).toContain("  - Active account: true");
-            expect(secondLogin.exitCode).toBe(0);
-            expect(authFileContent.split("[[auth]]").length - 1).toBe(1);
-            expect(authFileContent).toContain("id = \"user-1\"");
-            expect(authFileContent).toContain("api_key = \"secret-2\"");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports auth login with a custom OOMOL_ENDPOINT", async () => {
-        const sandbox = await createCliSandbox();
-
-        sandbox.env.OOMOL_ENDPOINT = "staging.oomol.test";
-
-        try {
-            const result = await runPrintedAuthLogin(sandbox, "secret-1", {
-                accountEndpoint: sandbox.env.OOMOL_ENDPOINT,
-            });
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toContain(
-                readAuthLoginUrlPrefix("staging.oomol.test"),
-            );
-            expect(result.stdout).toContain(
-                "✓ Logged in to staging.oomol.test account Alice",
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports login as an alias for auth login", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const result = await runPrintedAuthLogin(sandbox, "secret-1", {
-                argv: ["login"],
-            });
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toContain(
-                "Open this URL in your browser to continue:",
-            );
-            expect(result.stdout).toContain("✓ Logged in to oomol.com account Alice");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("renders the auth login url and success block with color styling when stdout supports colors", async () => {
-        const sandbox = await createCliSandbox();
-        const colors = createTerminalColors(true);
-
-        try {
-            const login = await runPrintedAuthLogin(sandbox, "secret-1", {
-                stdoutHasColors: true,
-            });
-            const plainLoginUrl = findLoginUrl(login.stdout);
-
-            expect(login.exitCode).toBe(0);
-            expect(plainLoginUrl).toBeTruthy();
-            expect(login.stdout).toContain(
-                colors.hex(loginUrlColor)(plainLoginUrl!),
-            );
-            expect(login.stdout).toContain(colors.green("✓"));
-            expect(login.stdout).toContain(colors.bold("oomol.com"));
-            expect(login.stdout).toContain(colors.bold("Alice"));
-            expect(login.stdout).toContain(colors.bold("true"));
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports auth logout without falling back to another account", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-2\"",
-                    "name = \"Bob\"",
-                    "api_key = \"secret-2\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(["auth", "logout"]);
-            const authFileContent = await readFile(authFilePath, "utf8");
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toContain("Logged out");
-            expect(authFileContent).toContain("id = \"\"");
-            expect(authFileContent).not.toContain("id = \"user-1\"");
-            expect(authFileContent).toContain("id = \"user-2\"");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports logout as an alias for auth logout", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(["logout"]);
-            const authFileContent = await readFile(authFilePath, "utf8");
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toContain("Logged out");
-            expect(authFileContent).toContain("id = \"\"");
-            expect(authFileContent).not.toContain("id = \"user-1\"");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports auth status for valid and invalid api keys", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-            const validRequests: Request[] = [];
-            const invalidRequests: Request[] = [];
-
-            const validStatus = await sandbox.run(
-                ["auth", "status"],
-                {
-                    fetcher: async (input, init) => {
-                        validRequests.push(toRequest(input, init));
-                        return new Response(null, { status: 200 });
-                    },
-                },
-            );
-            const invalidStatus = await sandbox.run(
-                ["auth", "status"],
-                {
-                    fetcher: async (input, init) => {
-                        invalidRequests.push(toRequest(input, init));
-                        return new Response(null, { status: 401 });
-                    },
-                },
-            );
-
-            expect(validStatus.exitCode).toBe(0);
-            expect(validStatus.stdout).toContain("✓ Logged in to oomol.com account Alice");
-            expect(validStatus.stdout).toContain("  - Active account: true");
-            expect(validStatus.stdout).toContain("  - API key status: Valid");
-            expect(validStatus.stdout).not.toContain("saved_accounts=");
-            expect(validRequests).toHaveLength(1);
-            expect(validRequests[0]?.url).toBe("https://api.oomol.com/v1/users/profile");
-            expect(validRequests[0]?.headers.get("Authorization")).toBe("secret-1");
-
-            expect(invalidStatus.exitCode).toBe(0);
-            expect(invalidStatus.stdout).toContain("X Logged in to oomol.com account Alice");
-            expect(invalidStatus.stdout).toContain("  - API key status: Invalid");
-            expect(invalidRequests).toHaveLength(1);
-            expect(invalidRequests[0]?.url).toBe("https://api.oomol.com/v1/users/profile");
-            expect(invalidRequests[0]?.headers.get("Authorization")).toBe("secret-1");
-            expect(await readFile(authFilePath, "utf8")).toContain("api_key = \"secret-1\"");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports auth switch by activating the next saved account", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-2\"",
-                    "name = \"Bob\"",
-                    "api_key = \"secret-2\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-3\"",
-                    "name = \"Charlie\"",
-                    "api_key = \"secret-3\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const firstResult = await sandbox.run(["auth", "switch"]);
-
-            expect(firstResult.exitCode).toBe(0);
-            expect(firstResult.stdout).toContain(
-                "✓ Switched active account for oomol.com to Bob",
-            );
-            expect(await readFile(authFilePath, "utf8")).toContain("id = \"user-2\"");
-
-            const secondResult = await sandbox.run(["auth", "switch"]);
-
-            expect(secondResult.exitCode).toBe(0);
-            expect(secondResult.stdout).toContain(
-                "✓ Switched active account for oomol.com to Charlie",
-            );
-            expect(await readFile(authFilePath, "utf8")).toContain("id = \"user-3\"");
-
-            const thirdResult = await sandbox.run(["auth", "switch"]);
-
-            expect(thirdResult.exitCode).toBe(0);
-            expect(thirdResult.stdout).toContain(
-                "✓ Switched active account for oomol.com to Alice",
-            );
-            expect(await readFile(authFilePath, "utf8")).toContain("id = \"user-1\"");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports search command with text output", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const requests: Request[] = [];
-            const result = await sandbox.run(
-                ["search", "image processing"],
-                {
-                    fetcher: async (input, init) => {
-                        requests.push(toRequest(input, init));
-
-                        return new Response(JSON.stringify({
-                            packages: [
-                                {
-                                    name: "@oomol/image-tools",
-                                    version: "1.2.3",
-                                    displayName: "Image Tools",
-                                    description: "Powerful image processing toolkit",
-                                    blocks: [
-                                        {
-                                            name: "image-processor",
-                                            title: "Image Processor",
-                                            description:
-                                                "Process and transform image formats",
-                                        },
-                                    ],
-                                },
-                            ],
-                        }));
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(result.stdout).toBe(
-                [
-                    "Image Tools (@oomol/image-tools@1.2.3)",
-                    "Powerful image processing toolkit",
-                    "Blocks:",
-                    "- Image Processor (image-processor)",
-                    "  Process and transform image formats",
-                    "",
-                ].join("\n"),
-            );
-            expect(requests).toHaveLength(1);
-            expect(requests[0]?.url).toBe(
-                "https://search.oomol.com/v1/packages/-/intent-search?q=image+processing&lang=en",
-            );
-            expect(requests[0]?.headers.get("Authorization")).toBe("secret-1");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("adds the localized request language to search queries", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const requests: Request[] = [];
-            const result = await sandbox.run(
-                ["--lang", "zh", "search", "image processing"],
-                {
-                    fetcher: async (input, init) => {
-                        requests.push(toRequest(input, init));
-
-                        return new Response(JSON.stringify({
-                            packages: [],
-                        }));
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(requests).toHaveLength(1);
-            expect(
-                new URL(requests[0]!.url).searchParams.get("lang"),
-            ).toBe("zh-CN");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("reuses cached search responses across cli invocations", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            let requestCount = 0;
-            const fetcher = async () => {
-                requestCount += 1;
-
-                return new Response(JSON.stringify({
-                    packages: [
-                        {
-                            name: "@oomol/image-tools",
-                            version: "1.2.3",
-                            displayName: "Image Tools",
-                            description: "Powerful image processing toolkit",
-                        },
-                    ],
-                }));
-            };
-            const firstResult = await sandbox.run(
-                ["search", "image processing"],
-                { fetcher },
-            );
-            const secondResult = await sandbox.run(
-                ["search", "image processing"],
-                { fetcher },
-            );
-
-            expect(firstResult.exitCode).toBe(0);
-            expect(secondResult.exitCode).toBe(0);
-            expect(firstResult.stdout).toBe(secondResult.stdout);
-            expect(requestCount).toBe(1);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("renders search output with field-specific colors", async () => {
-        const sandbox = await createCliSandbox();
-        const colors = createTerminalColors(true);
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(
-                ["search", "image processing"],
-                {
-                    fetcher: async () => new Response(JSON.stringify({
-                        packages: [
-                            {
-                                name: "@oomol/image-tools",
-                                version: "1.2.3",
-                                displayName: "Image Tools",
-                                description: "Powerful image processing toolkit",
-                                blocks: [
-                                    {
-                                        name: "image-processor",
-                                        title: "Image Processor",
-                                        description:
-                                            "Process and transform image formats",
-                                    },
-                                ],
-                            },
-                        ],
-                    })),
-                    stdout: {
-                        hasColors: true,
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(createTerminalColors(true).strip(result.stdout)).toBe(
-                [
-                    "Image Tools (@oomol/image-tools@1.2.3)",
-                    "Powerful image processing toolkit",
-                    "Blocks:",
-                    "- Image Processor (image-processor)",
-                    "  Process and transform image formats",
-                    "",
-                ].join("\n"),
-            );
-            expect(result.stdout).toContain(
-                `${colors.hex(searchDisplayNameColor)("Image Tools")} (@oomol/image-tools@1.2.3)`,
-            );
-            expect(result.stdout).toContain(
-                `${colors.hex(searchBlockTitleColor)("Image Processor")} (image-processor)`,
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports search command with only-package-id text output", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(
-                ["search", "image processing", "--only-package-id"],
-                {
-                    fetcher: async () => new Response(JSON.stringify({
-                        packages: [
-                            {
-                                name: "@oomol/image-tools",
-                                version: "1.2.3",
-                                displayName: "Image Tools",
-                                description: "Powerful image processing toolkit",
-                            },
-                            {
-                                name: "@oomol/vision-kit",
-                                version: "2.0.0",
-                                displayName: "Vision Kit",
-                            },
-                        ],
-                    })),
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(result.stdout).toBe(
-                [
-                    "@oomol/image-tools@1.2.3",
-                    "@oomol/vision-kit@2.0.0",
-                    "",
-                ].join("\n"),
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports search command with json array output and trims long text", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const rawResponse = JSON.stringify({
-                packages: [
-                    {
-                        blocks: [
-                            {
-                                title: "Image Processor",
-                            },
-                        ],
-                        displayName: "Image Tools",
-                        name: "@oomol/image-tools",
-                        version: "1.2.3",
-                    },
-                ],
-                total: 1,
-            });
-            const requests: Request[] = [];
-            const searchText = "x".repeat(210);
-            const expectedQuery = "x".repeat(200);
-            const result = await sandbox.run(
-                ["search", searchText, "--json"],
-                {
-                    fetcher: async (input, init) => {
-                        requests.push(toRequest(input, init));
-
-                        return new Response(rawResponse);
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toBe(`${JSON.stringify([
-                {
-                    blocks: [
-                        {
-                            title: "Image Processor",
-                        },
-                    ],
-                    displayName: "Image Tools",
-                    name: "@oomol/image-tools",
-                    version: "1.2.3",
-                },
-            ])}\n`);
-            expect(result.stderr).toBe("");
-            expect(requests).toHaveLength(1);
-            expect(
-                new URL(requests[0]!.url).searchParams.get("q"),
-            ).toBe(expectedQuery);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports search command with only-package-id json output", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(
-                ["search", "image processing", "--format=json", "--only-package-id"],
-                {
-                    fetcher: async () => new Response(JSON.stringify({
-                        packages: [
-                            {
-                                name: "@oomol/image-tools",
-                                version: "1.2.3",
-                                displayName: "Image Tools",
-                            },
-                            {
-                                name: "@oomol/vision-kit",
-                                version: "2.0.0",
-                            },
-                        ],
-                    })),
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(result.stdout).toBe(`${JSON.stringify([
-                "@oomol/image-tools@1.2.3",
-                "@oomol/vision-kit@2.0.0",
-            ])}\n`);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports package info command with text output", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const requests: Request[] = [];
-            const result = await sandbox.run(
-                ["package", "info", "qrcode"],
-                {
-                    fetcher: async (input, init) => {
-                        requests.push(toRequest(input, init));
-
-                        return new Response(JSON.stringify({
-                            packageName: "qrcode",
-                            packageVersion: "1.0.4",
-                            title: "QR Code",
-                            description: "The QR Code Toolkit.",
-                            blocks: [
-                                {
-                                    blockName: "Exist",
-                                    title: "Exist QR Code",
-                                    description: "Checks whether an image contains a QR code.",
-                                    inputHandleDefs: [
-                                        {
-                                            group: "Image Input",
-                                            collapsed: true,
-                                        },
-                                        {
-                                            handle: "input",
-                                            description: "Image input",
-                                            value: "sample.png",
-                                            json_schema: {
-                                                "contentMediaType": "oomol/image",
-                                                "anyOf": [
-                                                    {
-                                                        "type": "string",
-                                                        "ui:widget": "text",
-                                                    },
-                                                ],
-                                                "ui:options": {
-                                                    labels: ["Base64 with Text"],
-                                                },
-                                            },
-                                        },
-                                        {
-                                            handle: "tags",
-                                            description: "Tag list",
-                                            json_schema: {
-                                                type: "array",
-                                                items: {
-                                                    type: "string",
-                                                },
-                                            },
-                                        },
-                                        {
-                                            handle: "mode",
-                                            description: "Scan mode",
-                                            json_schema: {
-                                                type: "string",
-                                                default: "auto",
-                                            },
-                                        },
-                                        {
-                                            handle: "count",
-                                            description: "Retry count",
-                                            nullable: true,
-                                            value: null,
-                                            json_schema: {
-                                                type: "integer",
-                                            },
-                                        },
-                                    ],
-                                    outputHandleDefs: [
-                                        {
-                                            handle: "output",
-                                            description: "Boolean result",
-                                            json_schema: {
-                                                "type": "boolean",
-                                                "ui:widget": "switch",
-                                            },
-                                        },
-                                        {
-                                            handle: "metadata",
-                                            description: "Unstructured metadata",
-                                            json_schema: {},
-                                        },
-                                    ],
-                                },
-                                {
-                                    blockName: "Decode",
-                                    title: "Decode QR Code",
-                                    description: "Reads the QR code payload from an image.",
-                                    inputHandleDefs: [
-                                        {
-                                            handle: "image",
-                                            description: "Image to decode",
-                                            json_schema: {
-                                                type: "string",
-                                                contentMediaType: "oomol/image",
-                                            },
-                                        },
-                                    ],
-                                    outputHandleDefs: [
-                                        {
-                                            handle: "text",
-                                            description: "Decoded text payload",
-                                            json_schema: {
-                                                type: "string",
-                                            },
-                                        },
-                                    ],
-                                },
-                            ],
-                        }));
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(result.stdout).toBe(
-                [
-                    "QR Code (qrcode@1.0.4)",
-                    "The QR Code Toolkit.",
-                    "",
-                    "- Exist QR Code (Exist)",
-                    "  Checks whether an image contains a QR code.",
-                    "  Input:",
-                    "    - input  string (image)  [optional]  Image input",
-                    "    - tags   Array<string>   [required]  Tag list",
-                    "    - mode   string          [optional]  Scan mode",
-                    "    - count  integer         [optional]  Retry count",
-                    "  Output:",
-                    "    - output    boolean  Boolean result",
-                    "    - metadata  unknown  Unstructured metadata",
-                    "",
-                    "- Decode QR Code (Decode)",
-                    "  Reads the QR code payload from an image.",
-                    "  Input:",
-                    "    - image  string (image)  [required]  Image to decode",
-                    "  Output:",
-                    "    - text  string  Decoded text payload",
-                    "",
-                ].join("\n"),
-            );
-            expect(requests).toHaveLength(1);
-            expect(requests[0]?.url).toBe(
-                "https://registry.oomol.com/-/oomol/package-info/qrcode/latest?lang=en",
-            );
-            expect(requests[0]?.headers.get("Authorization")).toBe("secret-1");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("normalizes input ext metadata, applies input schema patches, and keeps output handle schema raw in package info json output", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(
-                ["package", "info", "qrcode@1.0.4", "--json"],
-                {
-                    fetcher: async () => new Response(JSON.stringify({
-                        packageName: "qrcode",
-                        packageVersion: "1.0.4",
-                        title: "QR Code",
-                        description: "The QR Code Toolkit.",
-                        blocks: [
-                            {
-                                blockName: "Exist",
-                                title: "Exist QR Code",
-                                description: "Checks whether an image contains a QR code.",
-                                inputHandleDefs: [
-                                    {
-                                        handle: "input",
-                                        description: "Image input",
-                                        nullable: false,
-                                        value: "sample.png",
-                                        json_schema: {
-                                            "anyOf": [
-                                                {
-                                                    "type": "string",
-                                                    "ui:widget": "text",
-                                                    "ui:placeholder": "Base64 text input",
-                                                },
-                                            ],
-                                            "ui:help": "ignored",
-                                            "ui:options": {
-                                                labels: ["Base64 with Text"],
-                                            },
-                                        },
-                                    },
-                                    {
-                                        handle: "placeholder",
-                                        description: "Optional placeholder",
-                                        nullable: true,
-                                        value: null,
-                                        json_schema: {
-                                            type: "string",
-                                        },
-                                    },
-                                    {
-                                        handle: "fileInput",
-                                        description: "Input file",
-                                        json_schema: {
-                                            "type": "string",
-                                            "ui:widget": "file",
-                                        },
-                                    },
-                                    {
-                                        handle: "excludes",
-                                        description: "Excluded usernames",
-                                        value: ["alice", "bob"],
-                                        json_schema: {
-                                            type: "array",
-                                            items: {
-                                                type: "string",
-                                            },
-                                        },
-                                    },
-                                    {
-                                        handle: "count",
-                                        description: "Winner count",
-                                        value: 3,
-                                        json_schema: {
-                                            type: "integer",
-                                        },
-                                    },
-                                    {
-                                        handle: "tags",
-                                        description: "Tag list",
-                                        json_schema: {
-                                            type: "array",
-                                            items: {
-                                                type: "string",
-                                            },
-                                        },
-                                    },
-                                ],
-                                outputHandleDefs: [
-                                    {
-                                        handle: "output",
-                                        description: "Boolean result",
-                                        json_schema: {
-                                            "type": "boolean",
-                                            "ui:widget": "switch",
-                                            "ui:tone": "success",
-                                        },
-                                    },
-                                ],
-                            },
-                        ],
-                    })),
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(JSON.parse(result.stdout)).toEqual({
-                blocks: [
-                    {
-                        blockName: "Exist",
-                        title: "Exist QR Code",
-                        description: "Checks whether an image contains a QR code.",
-                        inputHandle: {
-                            input: {
-                                description: "Image input",
-                                ext: {
-                                    anyOf: [
-                                        {
-                                            widget: "text",
-                                        },
-                                    ],
-                                },
-                                nullable: false,
-                                schema: {
-                                    anyOf: [
-                                        {
-                                            type: "string",
-                                        },
-                                    ],
-                                },
-                                value: "sample.png",
-                            },
-                            placeholder: {
-                                description: "Optional placeholder",
-                                nullable: true,
-                                schema: {
-                                    type: "string",
-                                },
-                                value: null,
-                            },
-                            fileInput: {
-                                description: "Input file",
-                                ext: {
-                                    widget: "file",
-                                },
-                                schema: {
-                                    format: "uri",
-                                    type: "string",
-                                },
-                            },
-                            excludes: {
-                                description: "Excluded usernames",
-                                schema: {
-                                    type: "array",
-                                    items: {
-                                        type: "string",
-                                    },
-                                },
-                                value: ["alice", "bob"],
-                            },
-                            count: {
-                                description: "Winner count",
-                                schema: {
-                                    type: "integer",
-                                },
-                                value: 3,
-                            },
-                            tags: {
-                                description: "Tag list",
-                                schema: {
-                                    type: "array",
-                                    items: {
-                                        type: "string",
-                                    },
-                                },
-                            },
-                        },
-                        outputHandle: {
-                            output: {
-                                description: "Boolean result",
-                                schema: {
-                                    "ui:tone": "success",
-                                    "ui:widget": "switch",
-                                    "type": "boolean",
-                                },
-                            },
-                        },
-                    },
-                ],
-                description: "The QR Code Toolkit.",
-                displayName: "QR Code",
-                packageName: "qrcode",
-                packageVersion: "1.0.4",
-            });
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports package info package specifier variants and json output", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const requests: Request[] = [];
-            const cases = [
-                {
-                    argv: ["package", "info", "pdf@1.0.0", "--format=json"],
-                    expectedUrl:
-                        "https://registry.oomol.com/-/oomol/package-info/pdf/1.0.0?lang=en",
-                    response: {
-                        packageName: "pdf",
-                        packageVersion: "1.0.0",
-                        title: "PDF Toolkit",
-                        description: "Inspect PDF files",
-                        blocks: [],
-                    },
-                },
-                {
-                    argv: ["package", "info", "pdf", "--format=json"],
-                    expectedUrl:
-                        "https://registry.oomol.com/-/oomol/package-info/pdf/latest?lang=en",
-                    response: {
-                        packageName: "pdf",
-                        packageVersion: "1.0.0",
-                        title: "PDF Toolkit",
-                        description: "Inspect PDF files",
-                        blocks: [],
-                    },
-                },
-                {
-                    argv: ["package", "info", "@foo/epub", "--format=json"],
-                    expectedUrl:
-                        "https://registry.oomol.com/-/oomol/package-info/%40foo%2Fepub/latest?lang=en",
-                    response: {
-                        packageName: "@foo/epub",
-                        packageVersion: "2.0.0",
-                        title: "Scoped EPUB",
-                        description: "Read EPUB packages",
-                        blocks: [],
-                    },
-                },
-                {
-                    argv: ["package", "info", "@bar/epub@1.0.0", "--format=json"],
-                    expectedUrl:
-                        "https://registry.oomol.com/-/oomol/package-info/%40bar%2Fepub/1.0.0?lang=en",
-                    response: {
-                        packageName: "@bar/epub",
-                        packageVersion: "1.0.0",
-                        title: "Bar EPUB",
-                        description: "Read EPUB packages",
-                        blocks: [],
-                    },
-                },
-                {
-                    argv: ["package", "info", "@baz@md@latest", "--format=json"],
-                    expectedUrl:
-                        "https://registry.oomol.com/-/oomol/package-info/%40baz%40md/latest?lang=en",
-                    response: {
-                        packageName: "@baz@md",
-                        packageVersion: "3.2.1",
-                        title: "Baz Markdown",
-                        description: "Read Markdown packages",
-                        blocks: [],
-                    },
-                },
-            ] as const;
-
-            for (const testCase of cases) {
-                const result = await sandbox.run(
-                    [...testCase.argv],
-                    {
-                        fetcher: async (input, init) => {
-                            requests.push(toRequest(input, init));
-
-                            return new Response(JSON.stringify(testCase.response));
-                        },
-                    },
-                );
-
-                expect(result.exitCode).toBe(0);
-                expect(result.stderr).toBe("");
-                expect(JSON.parse(result.stdout)).toEqual({
-                    blocks: [],
-                    description: testCase.response.description,
-                    displayName: testCase.response.title,
-                    packageName: testCase.response.packageName,
-                    packageVersion: testCase.response.packageVersion,
-                });
-            }
-
-            expect(requests.map(request => request.url)).toEqual(
-                cases.map(testCase => testCase.expectedUrl),
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports cloud-task run with json output", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const requests: Request[] = [];
-            const result = await sandbox.run(
-                [
-                    "cloud-task",
-                    "run",
-                    "qrcode@1.0.4",
-                    "-b",
-                    "Exist",
-                    "-d",
-                    "{\"count\":3}",
-                    "--json",
-                ],
-                {
-                    fetcher: async (input, init) => {
-                        const request = toRequest(input, init);
-
-                        requests.push(request);
-
-                        if (request.url.startsWith("https://registry.")) {
-                            return new Response(JSON.stringify({
-                                packageName: "qrcode",
-                                packageVersion: "1.0.4",
-                                title: "QR Code",
-                                description: "The QR Code Toolkit.",
-                                blocks: [
-                                    {
-                                        blockName: "Exist",
-                                        title: "Exist QR Code",
-                                        description: "Checks whether an image contains a QR code.",
-                                        inputHandleDefs: [
-                                            {
-                                                handle: "count",
-                                                description: "Retry count",
-                                                json_schema: {
-                                                    type: "integer",
-                                                },
-                                            },
-                                        ],
-                                        outputHandleDefs: [],
-                                    },
-                                ],
-                            }));
-                        }
-
-                        return new Response(JSON.stringify({
-                            taskID: "550e8400-e29b-41d4-a716-446655440017",
-                        }));
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(JSON.parse(result.stdout)).toEqual({
-                taskID: "550e8400-e29b-41d4-a716-446655440017",
-            });
-            expect(requests).toHaveLength(2);
-            expect(requests[0]?.url).toBe(
-                "https://registry.oomol.com/-/oomol/package-info/qrcode/1.0.4?lang=en",
-            );
-            expect(requests[1]?.url).toBe(
-                "https://cloud-task.oomol.com/v3/users/me/tasks",
-            );
-            expect(requests[1]?.method).toBe("POST");
-            expect(requests[1]?.headers.get("Authorization")).toBe("secret-1");
-            expect(requests[1]?.headers.get("Content-Type")).toBe("application/json");
-            await expect(requests[1]?.json()).resolves.toEqual({
-                blockName: "Exist",
-                inputValues: {
-                    count: 3,
-                },
-                packageName: "qrcode",
-                packageVersion: "1.0.4",
-                type: "serverless",
-            });
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports cloud-task run dry-run with @file payloads", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-            const payloadPath = join(sandbox.cwd, "cloud-task-payload.txt");
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-            await Bun.write(payloadPath, "{\"secret\":\"value\"}");
-
-            const requests: Request[] = [];
-            const result = await sandbox.run(
-                [
-                    "cloud-task",
-                    "run",
-                    "secret-tool@1.2.3",
-                    "-b",
-                    "main",
-                    "-d",
-                    `@${payloadPath}`,
-                    "--dry-run",
-                ],
-                {
-                    fetcher: async (input, init) => {
-                        const request = toRequest(input, init);
-
-                        requests.push(request);
-
-                        return new Response(JSON.stringify({
-                            packageName: "secret-tool",
-                            packageVersion: "1.2.3",
-                            title: "Secret Tool",
-                            description: "Handles secrets.",
-                            blocks: [
-                                {
-                                    blockName: "main",
-                                    title: "Main",
-                                    description: "Runs the main block.",
-                                    inputHandleDefs: [
-                                        {
-                                            handle: "secret",
-                                            description: "Secret value",
-                                            json_schema: {
-                                                contentMediaType: "oomol/secret",
-                                                type: "string",
-                                            },
-                                        },
-                                    ],
-                                    outputHandleDefs: [],
-                                },
-                            ],
-                        }));
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(result.stdout).toBe("Validation passed.\n");
-            expect(requests).toHaveLength(1);
-            expect(requests[0]?.url).toBe(
-                "https://registry.oomol.com/-/oomol/package-info/secret-tool/1.2.3?lang=en",
-            );
-        }
-        finally {
-            await Bun.file(join(sandbox.cwd, "cloud-task-payload.txt")).delete();
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports cloud-task result, log, and list json output", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const requests: Request[] = [];
-            const fetcher = async (input: string | URL | Request, init?: RequestInit) => {
-                const request = toRequest(input, init);
-
-                requests.push(request);
-
-                if (request.url.endsWith("/result")) {
-                    return new Response(JSON.stringify({
-                        resultData: { output: "ok" },
-                        resultURL: "https://example.com/result.json",
-                        status: "success",
-                        traceID: "trace-1",
-                    }));
-                }
-
-                if (request.url.includes("/logs")) {
-                    return new Response(JSON.stringify({
-                        logs: [
-                            {
-                                level: "info",
-                                message: "running",
-                            },
-                        ],
-                    }));
-                }
-
-                return new Response(JSON.stringify({
-                    nextToken: "eyJsYXN0SWQiOiIxMjMifQ==",
-                    tasks: [
-                        {
-                            blockName: "main",
-                            createdAt: 1704067200000,
-                            endTime: null,
-                            failedMessage: null,
-                            ownerID: "user-1",
-                            packageID: "foo",
-                            progress: 50,
-                            resultURL: null,
-                            schedulerPayload: {
-                                type: "serverless",
-                            },
-                            startTime: null,
-                            status: "running",
-                            subscriptionID: null,
-                            taskID: "550e8400-e29b-41d4-a716-446655440019",
-                            taskType: "user",
-                            updatedAt: 1704067200000,
-                            workload: "serverless",
-                            workloadID: "550e8400-e29b-41d4-a716-446655440020",
-                        },
-                    ],
-                }));
-            };
-
-            const resultResponse = await sandbox.run(
-                ["cloud-task", "result", "task-1", "--json"],
-                {
-                    fetcher,
-                },
-            );
-            const logResponse = await sandbox.run(
-                ["cloud-task", "log", "task-1", "--json", "--page=2"],
-                {
-                    fetcher,
-                },
-            );
-            const listResponse = await sandbox.run(
-                [
-                    "cloud-task",
-                    "list",
-                    "--json",
-                    "--size=1",
-                    "--nextToken=eyJsYXN0SWQiOiIxMjMifQ==",
-                    "--status=running",
-                    "--package-name=foo",
-                    "--block-name=main",
-                ],
-                {
-                    fetcher,
-                },
-            );
-
-            expect(resultResponse.exitCode).toBe(0);
-            expect(JSON.parse(resultResponse.stdout)).toEqual({
-                resultData: { output: "ok" },
-                resultURL: "https://example.com/result.json",
-                status: "success",
-                traceID: "trace-1",
-            });
-            expect(logResponse.exitCode).toBe(0);
-            expect(JSON.parse(logResponse.stdout)).toEqual({
-                logs: [
-                    {
-                        level: "info",
-                        message: "running",
-                    },
-                ],
-            });
-            expect(listResponse.exitCode).toBe(0);
-            expect(JSON.parse(listResponse.stdout)).toEqual({
-                nextToken: "eyJsYXN0SWQiOiIxMjMifQ==",
-                tasks: [
-                    {
-                        blockName: "main",
-                        createdAt: 1704067200000,
-                        endTime: null,
-                        failedMessage: null,
-                        ownerID: "user-1",
-                        packageID: "foo",
-                        progress: 50,
-                        resultURL: null,
-                        schedulerPayload: {
-                            type: "serverless",
-                        },
-                        startTime: null,
-                        status: "running",
-                        subscriptionID: null,
-                        taskID: "550e8400-e29b-41d4-a716-446655440019",
-                        taskType: "user",
-                        updatedAt: 1704067200000,
-                        workload: "serverless",
-                        workloadID: "550e8400-e29b-41d4-a716-446655440020",
-                    },
-                ],
-            });
-            expect(requests.map(request => request.url)).toEqual([
-                "https://cloud-task.oomol.com/v3/users/me/tasks/task-1/result",
-                "https://cloud-task.oomol.com/v3/users/me/tasks/task-1/logs?page=2",
-                "https://cloud-task.oomol.com/v3/users/me/tasks?size=1&nextToken=eyJsYXN0SWQiOiIxMjMifQ%3D%3D&status=running&packageID=foo&blockName=main",
-            ]);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports cloud-task wait and its alias with timeout parsing", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const requests: Request[] = [];
-            const fetcher = async (input: string | URL | Request, init?: RequestInit) => {
-                const request = toRequest(input, init);
-
-                requests.push(request);
-
-                return new Response(JSON.stringify({
-                    resultData: {
-                        output: "ok",
-                    },
-                    status: "success",
-                }));
-            };
-
-            const waitResponse = await sandbox.run(
-                ["cloud-task", "wait", "task-1", "--timeout=360"],
-                {
-                    fetcher,
-                },
-            );
-            const aliasResponse = await sandbox.run(
-                ["cloud-task", "wati", "task-2", "--timeout=1m"],
-                {
-                    fetcher,
-                },
-            );
-
-            expect(waitResponse.exitCode).toBe(0);
-            expect(waitResponse.stderr).toBe("");
-            expect(createTerminalColors(true).strip(waitResponse.stdout)).toBe(
-                [
-                    "✓ success",
-                    "  Task ID: task-1",
-                    "  Result data:",
-                    "    {",
-                    "      \"output\": \"ok\"",
-                    "    }",
-                    "",
-                ].join("\n"),
-            );
-            expect(aliasResponse.exitCode).toBe(0);
-            expect(aliasResponse.stderr).toBe("");
-            expect(createTerminalColors(true).strip(aliasResponse.stdout)).toBe(
-                [
-                    "✓ success",
-                    "  Task ID: task-2",
-                    "  Result data:",
-                    "    {",
-                    "      \"output\": \"ok\"",
-                    "    }",
-                    "",
-                ].join("\n"),
-            );
-            expect(requests.map(request => request.url)).toEqual([
-                "https://cloud-task.oomol.com/v3/users/me/tasks/task-1/result",
-                "https://cloud-task.oomol.com/v3/users/me/tasks/task-2/result",
-            ]);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("validates cloud-task wait timeout values", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            let fetchCount = 0;
-            const result = await sandbox.run(
-                ["cloud-task", "wait", "task-1", "--timeout=9s"],
-                {
-                    fetcher: async () => {
-                        fetchCount += 1;
-
-                        return new Response(JSON.stringify({
-                            status: "success",
-                        }));
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(2);
-            expect(result.stdout).toBe("");
-            expect(result.stderr).toBe(
-                "Invalid value for --timeout: 9s. Use 10s to 24h, with optional s, m, or h suffixes.\n",
-            );
-            expect(fetchCount).toBe(0);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("treats omitted input handles with default values as optional in cloud-task run", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const requests: Request[] = [];
-            const result = await sandbox.run(
-                [
-                    "cloud-task",
-                    "run",
-                    "hash-tool@1.0.0",
-                    "-b",
-                    "main",
-                    "-d",
-                    "{}",
-                    "--dry-run",
-                ],
-                {
-                    fetcher: async (input, init) => {
-                        requests.push(toRequest(input, init));
-
-                        return new Response(JSON.stringify({
-                            packageName: "hash-tool",
-                            packageVersion: "1.0.0",
-                            title: "Hash Tool",
-                            description: "Generate hash values.",
-                            blocks: [
-                                {
-                                    blockName: "main",
-                                    title: "Main",
-                                    description: "Generate hash values from input text.",
-                                    inputHandleDefs: [
-                                        {
-                                            handle: "input",
-                                            description: "Input text",
-                                            value: "",
-                                            json_schema: {
-                                                type: "string",
-                                            },
-                                        },
-                                        {
-                                            handle: "lowercased",
-                                            description: "Lowercase output",
-                                            value: false,
-                                            json_schema: {
-                                                type: "boolean",
-                                            },
-                                        },
-                                    ],
-                                    outputHandleDefs: [],
-                                },
-                            ],
-                        }));
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(result.stdout).toBe("Validation passed.\n");
-            expect(requests).toHaveLength(1);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("accepts omitted and empty --data values by treating them as empty objects", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const fetcher = async () => new Response(JSON.stringify({
-                packageName: "hash-tool",
-                packageVersion: "1.0.0",
-                title: "Hash Tool",
-                description: "Generate hash values.",
-                blocks: [
-                    {
-                        blockName: "main",
-                        title: "Main",
-                        description: "Generate hash values from input text.",
-                        inputHandleDefs: [
-                            {
-                                handle: "input",
-                                description: "Input text",
-                                value: "",
-                                json_schema: {
-                                    type: "string",
-                                },
-                            },
-                            {
-                                handle: "lowercased",
-                                description: "Lowercase output",
-                                value: false,
-                                json_schema: {
-                                    type: "boolean",
-                                },
-                            },
-                        ],
-                        outputHandleDefs: [],
-                    },
-                ],
-            }));
-
-            const omittedDataResult = await sandbox.run(
-                [
-                    "cloud-task",
-                    "run",
-                    "hash-tool@1.0.0",
-                    "-b",
-                    "main",
-                    "--dry-run",
-                ],
-                {
-                    fetcher,
-                },
-            );
-            const emptyDataResult = await sandbox.run(
-                [
-                    "cloud-task",
-                    "run",
-                    "hash-tool@1.0.0",
-                    "-b",
-                    "main",
-                    "-d",
-                    "",
-                    "--dry-run",
-                ],
-                {
-                    fetcher,
-                },
-            );
-
-            expect(omittedDataResult.exitCode).toBe(0);
-            expect(omittedDataResult.stderr).toBe("");
-            expect(omittedDataResult.stdout).toBe("Validation passed.\n");
-            expect(emptyDataResult.exitCode).toBe(0);
-            expect(emptyDataResult.stderr).toBe("");
-            expect(emptyDataResult.stdout).toBe("Validation passed.\n");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("renders cloud-task result text output with colors and grouped details", async () => {
-        const sandbox = await createCliSandbox();
-        const colors = createTerminalColors(true);
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(
-                ["cloud-task", "result", "task-1"],
-                {
-                    fetcher: async () => new Response(JSON.stringify({
-                        resultData: {
-                            output: "ok",
-                        },
-                        resultURL: "https://example.com/result.json",
-                        status: "success",
-                    })),
-                    stdout: {
-                        hasColors: true,
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(createTerminalColors(true).strip(result.stdout)).toBe(
-                [
-                    "✓ success",
-                    "  Task ID: task-1",
-                    "  Result URL: https://example.com/result.json",
-                    "  Result data:",
-                    "    {",
-                    "      \"output\": \"ok\"",
-                    "    }",
-                    "",
-                ].join("\n"),
-            );
-            expect(result.stdout).toContain(colors.green("✓"));
-            expect(result.stdout).toContain(colors.green.bold("success"));
-            expect(result.stdout).toContain(colors.bold("task-1"));
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("renders cloud-task list text output with colors and summary blocks", async () => {
-        const sandbox = await createCliSandbox();
-        const colors = createTerminalColors(true);
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(
-                ["cloud-task", "list", "--size=1"],
-                {
-                    fetcher: async () => new Response(JSON.stringify({
-                        nextToken: "next-token",
-                        tasks: [
-                            {
-                                blockName: "main",
-                                createdAt: 1704067200000,
-                                endTime: null,
-                                failedMessage: null,
-                                ownerID: "user-1",
-                                packageID: "foo",
-                                progress: 50,
-                                resultURL: null,
-                                schedulerPayload: {
-                                    inputValues: {
-                                        foo: "bar",
-                                    },
-                                    type: "serverless",
-                                },
-                                startTime: null,
-                                status: "running",
-                                subscriptionID: null,
-                                taskID: "task-1",
-                                taskType: "user",
-                                updatedAt: 1704067200000,
-                                workload: "serverless",
-                                workloadID: "workload-1",
-                            },
-                        ],
-                    })),
-                    stdout: {
-                        hasColors: true,
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(createTerminalColors(true).strip(result.stdout)).toBe(
-                [
-                    "▶ running",
-                    "  Task ID: task-1",
-                    "  Package/Block: foo / main",
-                    "  Workload: serverless",
-                    "  Progress: [=====-----] 50%",
-                    "  Created: 2024-01-01T00:00:00.000Z",
-                    "  Updated: 2024-01-01T00:00:00.000Z",
-                    "  Input values:",
-                    "    {",
-                    "      \"foo\": \"bar\"",
-                    "    }",
-                    "",
-                    "  Next token: next-token",
-                    "",
-                ].join("\n"),
-            );
-            expect(result.stdout).toContain(colors.blue("▶"));
-            expect(result.stdout).toContain(colors.blue.bold("running"));
-            expect(result.stdout).toContain(colors.hex(searchDisplayNameColor)("foo"));
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("requires a package filter when cloud-task list receives a block filter", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(["cloud-task", "list", "--block-id=main"]);
-
-            expect(result.exitCode).toBe(2);
-            expect(result.stdout).toBe("");
-            expect(result.stderr).toBe(
-                "You must provide --package-id (or --package-name) when using --block-id.\n",
-            );
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("reuses cached package info responses for explicit versions", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            let requestCount = 0;
-            const fetcher = async () => {
-                requestCount += 1;
-
-                return new Response(JSON.stringify({
-                    packageName: "qrcode",
-                    packageVersion: "1.0.4",
-                    title: "QR Code",
-                    description: "The QR Code Toolkit.",
-                    blocks: [],
-                }));
-            };
-
-            const firstResult = await sandbox.run(
-                ["package", "info", "qrcode@1.0.4"],
-                { fetcher },
-            );
-            const secondResult = await sandbox.run(
-                ["package", "info", "qrcode@1.0.4"],
-                { fetcher },
-            );
-
-            expect(firstResult.exitCode).toBe(0);
-            expect(secondResult.exitCode).toBe(0);
-            expect(firstResult.stdout).toBe(secondResult.stdout);
-            expect(requestCount).toBe(1);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("does not read latest package info lookups from cache and backfills the resolved version", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            let requestCount = 0;
-            const fetcher = async () => {
-                requestCount += 1;
-
-                return new Response(JSON.stringify({
-                    packageName: "qrcode",
-                    packageVersion: "1.0.4",
-                    title: "QR Code",
-                    description: "The QR Code Toolkit.",
-                    blocks: [],
-                }));
-            };
-
-            const latestResult = await sandbox.run(
-                ["package", "info", "qrcode"],
-                { fetcher },
-            );
-            const latestAgainResult = await sandbox.run(
-                ["package", "info", "qrcode"],
-                { fetcher },
-            );
-            const explicitVersionResult = await sandbox.run(
-                ["package", "info", "qrcode@1.0.4"],
-                { fetcher },
-            );
-
-            expect(latestResult.exitCode).toBe(0);
-            expect(latestAgainResult.exitCode).toBe(0);
-            expect(explicitVersionResult.exitCode).toBe(0);
-            expect(requestCount).toBe(2);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("validates the search format option", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const result = await sandbox.run(["search", "image", "--format=yaml"]);
-
-            expect(result.exitCode).toBe(2);
-            expect(result.stdout).toBe("");
-            expect(result.stderr).toContain("Invalid format: yaml. Use json.");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports package info command help with the --json alias", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const result = await sandbox.run(["package", "info", "--help"]);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(result.stdout).toContain("--json");
-            expect(result.stdout).toContain("Alias for --format=json");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("supports file upload command help with the --json alias", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const result = await sandbox.run(["file", "upload", "--help"]);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(result.stdout).toContain("--json");
-            expect(result.stdout).toContain("Alias for --format=json");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("renders search help when text argument is omitted", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const expectedHelp = await sandbox.run(["search", "--help"]);
-            const result = await sandbox.run(["search"]);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stderr).toBe("");
-            expect(result.stdout).toBe(expectedHelp.stdout);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("renders the auth switch success block with gh-style emphasis when stdout supports colors", async () => {
-        const sandbox = await createCliSandbox();
-        const colors = createTerminalColors(true);
-
-        try {
-            const authFilePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "auth.toml",
-            );
-
-            await Bun.write(
-                authFilePath,
-                [
-                    "id = \"user-1\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-1\"",
-                    "name = \"Alice\"",
-                    "api_key = \"secret-1\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                    "[[auth]]",
-                    "id = \"user-2\"",
-                    "name = \"Bob\"",
-                    "api_key = \"secret-2\"",
-                    "endpoint = \"oomol.com\"",
-                    "",
-                ].join("\n"),
-            );
-
-            const result = await sandbox.run(
-                ["auth", "switch"],
-                {
-                    stdout: {
-                        hasColors: true,
-                    },
-                },
-            );
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toContain(colors.green("✓"));
-            expect(result.stdout).toContain(
-                "Switched active account for oomol.com to",
-            );
-            expect(result.stdout).toContain(colors.bold("Bob"));
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("creates a default settings file on first read", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const filePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "settings.toml",
-            );
-
-            const result = await sandbox.run(["config", "list"]);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toBe("");
-            expect(result.stderr).toBe("");
-            expect(await readFile(filePath, "utf8")).toBe(defaultSettingsFileContent);
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("returns usage errors for invalid inputs", async () => {
+    test("returns usage errors for invalid global inputs", async () => {
         const sandbox = await createCliSandbox();
 
         try {
             const invalidLang = await sandbox.run(["--lang", "fr", "--help"]);
-            const invalidKey = await sandbox.run(["config", "get", "update-notifier"]);
-            const invalidConfigValue = await sandbox.run(["config", "set", "lang", "fr"]);
-            const invalidSkillPolicyValue = await sandbox.run([
-                "config",
-                "set",
-                "skills.oo.allow_implicit_invocation",
-                "maybe",
-            ]);
             const unknownCommand = await sandbox.run(["cnfig"]);
 
             expect(invalidLang.exitCode).toBe(2);
             expect(invalidLang.stderr).toContain("Invalid value for --lang");
 
-            expect(invalidKey.exitCode).toBe(2);
-            expect(invalidKey.stderr).toContain("Invalid config key");
-            expect(invalidKey.stderr).not.toContain("Supported keys");
-
-            expect(invalidConfigValue.exitCode).toBe(2);
-            expect(invalidConfigValue.stderr).toContain("Invalid lang value");
-
-            expect(invalidSkillPolicyValue.exitCode).toBe(2);
-            expect(invalidSkillPolicyValue.stderr).toContain(
-                "Invalid skills.oo.allow_implicit_invocation value",
-            );
-
             expect(unknownCommand.exitCode).toBe(2);
             expect(unknownCommand.stderr).toContain("Unknown command");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
-
-    test("returns runtime errors when the persisted store is corrupted", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const filePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "settings.toml",
-            );
-
-            await Bun.write(filePath, "{");
-
-            const result = await sandbox.run(["config", "get", "lang"]);
-            const content = await readLatestLogContent(sandbox);
-
-            expect(result.exitCode).toBe(1);
-            expect(result.stderr).toContain("settings file");
-            expect(content).toContain(`"category":"system_error"`);
-            expect(content).toContain(`"key":"errors.store.invalidToml"`);
         }
         finally {
             await sandbox.cleanup();
@@ -5113,206 +255,4 @@ describe("runCli", () => {
             await sandbox.cleanup();
         }
     });
-
-    test("reads TOML settings files", async () => {
-        const sandbox = await createCliSandbox();
-
-        try {
-            const filePath = join(
-                sandbox.env.XDG_CONFIG_HOME!,
-                APP_NAME,
-                "settings.toml",
-            );
-
-            await Bun.write(
-                filePath,
-                "lang = \"zh\"\n",
-            );
-
-            const result = await sandbox.run(["config", "get", "lang"]);
-
-            expect(result.exitCode).toBe(0);
-            expect(result.stdout).toBe("zh\n");
-            expect(result.stderr).toBe("");
-        }
-        finally {
-            await sandbox.cleanup();
-        }
-    });
 });
-
-async function runPrintedAuthLogin(
-    sandbox: Awaited<ReturnType<typeof createCliSandbox>>,
-    apiKeyValue: string,
-    options: {
-        accountEndpoint?: string;
-        argv?: readonly string[];
-        stdoutHasColors?: boolean;
-    } = {},
-): Promise<{
-    exitCode: number;
-    stdout: string;
-    stderr: string;
-}> {
-    const stdout = createTextBuffer({
-        hasColors: options.stdoutHasColors,
-    });
-    const stderr = createTextBuffer();
-    const execution = executeCli({
-        argv: options.argv ?? ["auth", "login"],
-        cwd: sandbox.cwd,
-        env: sandbox.env,
-        stdout: stdout.writer,
-        stderr: stderr.writer,
-        systemLocale: "en-US",
-    });
-    const loginUrl = await waitForLoginUrl(stdout);
-
-    await completeLoginCallback(
-        loginUrl,
-        apiKeyValue,
-        options.accountEndpoint ?? defaultAuthEndpoint,
-    );
-
-    return {
-        exitCode: await execution,
-        stdout: stdout.read(),
-        stderr: stderr.read(),
-    };
-}
-
-async function readLatestLogContent(
-    sandbox: Awaited<ReturnType<typeof createCliSandbox>>,
-): Promise<string> {
-    const logDirectoryPath = resolveStorePaths({
-        appName: APP_NAME,
-        env: sandbox.env,
-        platform: process.platform,
-    }).logDirectoryPath;
-    const logFileNames = await readdir(logDirectoryPath);
-    const logFilesWithMetadata = await Promise.all(
-        logFileNames.map(async fileName => ({
-            fileName,
-            metadata: await stat(join(logDirectoryPath, fileName)),
-        })),
-    );
-    const latestLogFileName = logFilesWithMetadata
-        .sort((left, right) => {
-            const modifiedAtDelta
-                = left.metadata.mtimeMs - right.metadata.mtimeMs;
-
-            if (modifiedAtDelta !== 0) {
-                return modifiedAtDelta;
-            }
-
-            return left.fileName.localeCompare(right.fileName);
-        })
-        .at(-1)
-        ?.fileName;
-
-    if (!latestLogFileName) {
-        throw new Error("Expected at least one log file.");
-    }
-
-    return await readFile(join(logDirectoryPath, latestLogFileName), "utf8");
-}
-
-function countDownloadResumeSessions(downloadSessionsFilePath: string): number {
-    const database = new Database(downloadSessionsFilePath, {
-        strict: true,
-    });
-
-    try {
-        const row = database.query(
-            `SELECT COUNT(*) AS count FROM ${downloadResumeSessionsTableName}`,
-        ).get() as {
-            count: number;
-        };
-
-        return row.count;
-    }
-    finally {
-        database.close();
-    }
-}
-
-function readFileDownloadSuccessOutput(path: string): string {
-    return `Saved to: ${path}\n`;
-}
-
-async function waitForLoginUrl(
-    stdout: ReturnType<typeof createTextBuffer>,
-): Promise<string> {
-    const deadline = Date.now() + 1000;
-
-    while (Date.now() < deadline) {
-        const loginUrl = findLoginUrl(stdout.read());
-
-        if (loginUrl !== undefined) {
-            return loginUrl;
-        }
-
-        await Bun.sleep(10);
-    }
-
-    throw new Error("Timed out waiting for the printed login URL.");
-}
-
-function findLoginUrl(output: string): string | undefined {
-    const plainOutput = createTerminalColors(true).strip(output);
-
-    for (const line of plainOutput.split("\n")) {
-        const urlStart = line.indexOf("https://");
-
-        if (urlStart < 0) {
-            continue;
-        }
-
-        const candidate = line.slice(urlStart).trim();
-
-        if (candidate.includes("/v1/auth/redirect?")) {
-            return candidate;
-        }
-    }
-
-    return undefined;
-}
-
-async function completeLoginCallback(
-    loginUrlValue: string,
-    apiKeyValue: string,
-    endpoint: string,
-): Promise<void> {
-    const loginUrl = new URL(loginUrlValue);
-
-    expect(loginUrl.searchParams.get("cli_login")).toBe("true");
-
-    const redirectUrl = loginUrl.searchParams.get("redirect");
-
-    expect(redirectUrl).toBeTruthy();
-
-    const callbackUrl = new URL(redirectUrl!);
-    const requestUrl = new URL(callbackUrl.toString());
-    const encodedApiKey = Buffer.from(apiKeyValue, "utf8").toString("base64");
-
-    requestUrl.searchParams.set("apiKey", encodedApiKey);
-    requestUrl.searchParams.set("name", "Alice");
-    requestUrl.searchParams.set("endpoint", endpoint);
-    requestUrl.searchParams.set("id", "user-1");
-
-    const response = await fetch(requestUrl);
-
-    expect(response.status).toBe(200);
-}
-
-function readAuthLoginUrlPrefix(endpoint: string): string {
-    return `https://api.${endpoint}/v1/auth/redirect?`;
-}
-
-function toRequest(input: string | URL | Request, init?: RequestInit): Request {
-    if (input instanceof Request) {
-        return new Request(input, init);
-    }
-
-    return new Request(String(input), init);
-}
